% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{add_loo_criterion}
\alias{add_loo_criterion}
\title{Add model fit loo criterion to model objects}
\usage{
add_loo_criterion(model, n_params, ...)
}
\arguments{
\item{model}{brmsfit object}

\item{n_params}{numeric value. The number of params being estimated by the
model. i.e. how many parameters have non-constant priors? in the case of dose
response cureves, are all 4 parameters being predicted by the model?}

\item{...}{extra arguments passed on to `r brms::add_criterion``}
}
\value{
brmsfit object with loo criterion.
}
\description{
This is required to use loo compare to compare multiple model fits against
each other.
}
