% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-basic-stats.R
\name{basic_stats}
\alias{basic_stats}
\title{Give Basic Statistical Information About the Model Results}
\usage{
basic_stats(
  model,
  exclude_vars = c("__$", "lprior"),
  l_ci = 0.025,
  u_ci = 0.975
)
}
\arguments{
\item{model}{\link[brms:brmsfit-class]{brms::brmsfit} object}

\item{exclude_vars}{list of regular expressions to match what variables to
exclude.}

\item{l_ci}{\code{numeric} unit of the lower confidence interval}

\item{u_ci}{\code{numeric} unit of the upper confidence interval}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} object.
}
\description{
Return a \code{data.frame} containing summary statistics of
\code{bpfit} model. The summary statistics included are mean, median,
standard deviation, lower confidence interval, and upper confidence
interval.
}
\examples{
\dontrun{
  BayesPharma::basic_stats(
    model = my_sigmoid_model,
    l_ci = 0.025,
    u_ci = 0.975)
}

}
