% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPharma-bpfit.R
\docType{class}
\name{bpfit-class}
\alias{bpfit-class}
\alias{bpfit}
\title{Class \code{bpfit} of Models Fit with the \link{BayesPharma} Package}
\description{
Models fit with the \link{BayesPharma} package model function
represented as a \code{bpfit} object, which is a wrapper around the
\link[brms:brmsfit-class]{brms::brmsfit} that contains the posterior draws (samples), model formula,
Stan code, relevant data, and other information. Additionally, \code{bpfit}
objects contain a \code{bayes_pharma_info} list with the following elements:
}
\details{
\itemize{
\item \strong{model_type}: \code{character} for the type of model
\item elements from defined in the formula, including the treatment and response
variables and their units
}

The BayesPharma information is used by some of the plots and model analysis
methods.

Additionally, by default \code{bpfit} objects have their methods exposed by
default so that it is possible e.g. compare models with
\link[brms:loo_compare.brmsfit]{brms::loo_compare}.
}
