% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-log-dose.R
\name{calculate_log_dose}
\alias{calculate_log_dose}
\title{Add a log dose (in base 10) column to the input data.frame}
\usage{
calculate_log_dose(data, dose_col = "dose", molar_concentration = 1)
}
\arguments{
\item{data}{\code{data.frame} containing a column representing a dose in
molar units.}

\item{dose_col}{\code{expression} for dose column in the input
\code{data.frame} (Default: \code{"dose"})}

\item{molar_concentration}{\code{numeric} units of molar concentration of the
dose column (Default: 1).}
}
\value{
input \code{data.frame} with an additional \code{log_dose} column
which is the \verb{log10(<dose_col>) * molar_concentration}.
}
\description{
Given the dose as a column in a \code{data.frame} with a given
molar concentration add a new column of the log base-10 dose,
\code{log_dose}, in the \code{data.frame} and return it.
}
\examples{
\dontrun{
# Consider observations at doses of 1 μM and 0.1 μM.
# If the doses are given in molar units e.g. M then,
data <- data.frame(dose = c(1e-6, 1e-7)) |>
   BayesPharma::calculate_log_dose(dose_col = dose, molar_concentration = 1)
data$log_dose == c(-6, -7)

# If the doses are given with in units of nanomolar units e.g. nM then
data <- data.frame(dose_nM = c(1000, 100)) |>
  BayesPharma::calculate_log_dose(
    dose_col = dose_nM,
    molar_concentration = 1e-9)
data$log_dose == c(-6, -7)
}
}
