% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{cli_init_output_path}
\alias{cli_init_output_path}
\title{Helper function to initialize output for the BayesPharma CLI}
\usage{
cli_init_output_path(options)
}
\arguments{
\item{options}{\code{list} of command line arguments as parsed by
\code{\link[=cli_parse_options]{cli_parse_options()}}.}
}
\description{
check if the output directory exists and create it if it does
not
}
