% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{cli_load_bivariate_formula}
\alias{cli_load_bivariate_formula}
\title{Helper function to load sigmoid formula the BayesPharma CLI}
\usage{
cli_load_bivariate_formula(options)
}
\arguments{
\item{options}{\code{list} of command line arguments as parsed by
\code{\link[=cli_parse_options]{cli_parse_options()}}.}
}
\value{
\code{BayesPharma::bpformula}
}
\description{
Initialize a bivaraite formula based on the command line
arguments. For the some models support multiple formulas, otherwise use the
default one for the model.
}
