% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{cli_load_data}
\alias{cli_load_data}
\title{Helper function to load data for the BayesPharma CLI}
\usage{
cli_load_data(options)
}
\arguments{
\item{options}{\code{list} of command line arguments as parsed by
\code{\link[=cli_parse_options]{cli_parse_options()}}.}
}
\value{
\code{data.frame}
}
\description{
load tabular data based on the extension used to fit the model
Recognized data formats are .csv, .tsv, and .xlsx
}
