% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{cli_load_formula}
\alias{cli_load_formula}
\title{Helper function to load formula the BayesPharma CLI}
\usage{
cli_load_formula(options)
}
\arguments{
\item{options}{\code{list} of command line arguments as parsed by
\code{\link[=cli_parse_options]{cli_parse_options()}}.}
}
\value{
\code{BayesPharma::bpformula}
}
\description{
Initialize formula based on the command line arguments. For the
some models support multiple formulas, otherwise use the default one for
the model.
}
