% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{cli_load_init}
\alias{cli_load_init}
\title{Helper function to load init the BayesPharma CLI}
\usage{
cli_load_init(options)
}
\arguments{
\item{options}{\code{list} of command line arguments as parsed by
\code{\link[=cli_parse_options]{cli_parse_options()}}.}
}
\value{
\code{function}
}
\description{
Initialize init based on the command line arguments.
}
