% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{cli_load_model_fn}
\alias{cli_load_model_fn}
\title{Helper function to load model function the BayesPharma CLI}
\usage{
cli_load_model_fn(options)
}
\arguments{
\item{options}{\code{list} of command line arguments as parsed by
\code{\link[=cli_parse_options]{cli_parse_options()}}.}
}
\value{
\code{function} used to fit the requested model
}
\description{
Initialize model function based on the command line arguments
Available models are those defined in \verb{BayesPharma::<model>_model}
}
