% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{cli_load_prior}
\alias{cli_load_prior}
\title{Helper function to load prior the BayesPharma CLI}
\usage{
cli_load_prior(options)
}
\arguments{
\item{options}{\code{list} of command line arguments as parsed by
\code{\link[=cli_parse_options]{cli_parse_options()}}.}
}
\value{
\code{brms::brmsprior}
}
\description{
Initialize prior based on the command line arguments.
}
