% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{compare_models}
\alias{compare_models}
\title{Compare different model fits against each other using loo compare}
\usage{
compare_models(model1, model2, ...)
}
\arguments{
\item{model1}{brmsfit object with loo criterion.}

\item{model2}{brmsfit object with loo criterion.}
}
\value{
compare.loo object
}
\description{
Use the add_loo_criterion function on the models that will be compared before
using this function to compare the models.
}
