% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density_distirbutions.R
\name{density_distributions}
\alias{density_distributions}
\title{Create a plot of the density distributions of modeled parameters from brmsfit
model}
\usage{
density_distributions(
  model,
  predictors_col_name = "_Intercept",
  half_max_label = "ec50",
  title_label = "Density Distributions",
  sample_type = "Prior"
)
}
\arguments{
\item{model}{brmsfit model.}

\item{predictors_col_name}{string expression for predictors column in the
input data.frame (default = "_Intercept"). Predictors are the perturbations
tested during the experiment (i.e. Drug, Temperature, etc.).}

\item{half_max_label}{string of the label for the half maximal that fits the
type of experiment that was done (i.e. ec50, ic50, ed50, id50, ld50, etc.).}

\item{title_label}{string of the plot title. (Default = "Prior Density Plot")}

\item{sample_type}{string of the type of density distribution
(i.e. Prior or Posterior). (default = "Prior")}
}
\value{
ggplot2::ggplot object.
}
\description{
This function is useful to observe the distributions of the
  priors set for the brmsfit model. By adding `sample_prior = "only"` to the
  dr_model or a brmsfit model, the model samples only from the prior
  distributions and can be observed using this plot function to verify that
  the values intended to be covered are being included.
}
\examples{
\dontrun{
  density_distributions(
    model = my_dr_model_priors,
    predictors_col_name = "predictors",
    half_max_label = "ic50",
    title_label = "Parameter Density Distribution Plots",
    sample_type = "Prior")
}
}
