% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density-distirbutions.R
\name{density_distributions_plot}
\alias{density_distributions_plot}
\title{Create a plot of the density distributions of modeled parameters from brmsfit
model}
\usage{
density_distributions_plot(
  model,
  pars = NULL,
  labeller = NULL,
  title_label = "Density Distributions"
)
}
\arguments{
\item{model}{brmsfit model.}

\item{pars}{parameters to choose (Default b_<parameter>_Intercept')}

\item{labeller}{strip off the 'b_' and '_Intercept' from the parameter labels}

\item{title_label}{string of the plot title. (Default = "Prior Density Plot")}
}
\value{
ggplot2::ggplot object.
}
\description{
This function is useful to observe the distributions
    of the priors set for the brmsfit model. By adding
    \code{sample_prior = "only"} to the sigmoid_model or a brmsfit
    model, the model samples only from the prior distributions and
    can be observed using this plot function to verify that the
    values intended to be covered are being included.
}
\examples{
\dontrun{
  density_distributions_plot(
    model = my_sigmoid_model)
}

}
