% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dr_brm_model.R
\name{dose_response_model}
\alias{dose_response_model}
\title{Run Bayesian Regression Model using Stan}
\usage{
dose_response_model(
  data,
  response_col_name,
  log_dose_col_name,
  predictors_col_name = NULL,
  sigmoid_formula = no_predictors_dr_formula(),
  priors = NULL,
  inits = 0,
  iter = 8000,
  control = list(adapt_delta = 0.99),
  ...
)
}
\arguments{
\item{data}{tibble or data.frame of experimental data.}

\item{priors}{brmspriors data.frame for ec50, hill, top, and bottom.
Use 'dr_priors' to create priors to use here.}

\item{inits}{list of lists, numeric value, or "random" for the initial values
of the parameters being modeled. (default = 0)}

\item{iter}{number of iterations the model runs. Increasing iter can help with
model convergence. (default = 8000)}

\item{control}{a named list of parameters to control the sampler's behavior.
Adding max_treedepth and giving a greater value than 10 can improve model
convergence. (default = list(adapt_delta = 0.99))}
}
\value{
brmsfit
}
\description{
For additional information on additional function arguments, reference:
https://paul-buerkner.github.io/brms/reference/brm.html
or
https://rdrr.io/cran/rstan/man/stan.html
}
