% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_inits.R
\name{dr_inits}
\alias{dr_inits}
\title{Create a list of initial values of the model parameters}
\usage{

}
\arguments{
\item{ec50}{numeric unit of ec50 (default = -9)}

\item{hill}{TRUE, FALSE, or numeric value, TRUE/FALSE sets hill to an
initial value of 1/-1 or a numeric value can be assigned to hill.
numeric unit of hill (default = TRUE)}

\item{inhibitor}{TRUE/FALSE value that determines if hill is a positive or
negative slope. If TRUE, the initial condition will be hill = -1. If FALSE,
hill = 1 (default = TRUE)}

\item{top}{numeric unit of top (default = 100)}

\item{bottom}{numeric unit of bottom (default = 0)}

\item{chain}{the number of Markov chains that will be used in brmsfit.(default = 4)}
}
\value{
list of lists
}
\description{
Creating a list of lists containing initial values for the model parameters
  that will be used as an argument in the dr_model.
}
