% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-drc.R
\name{drc_model}
\alias{drc_model}
\title{Wrapper for DRC dose response models}
\usage{
drc_model(
  data,
  formula = response ~ log_dose,
  fct = drc::L.4(fixed = c(NA, NA, NA, NA), names = c("hill", "bottom", "top", "ec50")),
  ...
)
}
\arguments{
\item{data}{\code{data.frame}}

\item{formula}{\code{formula} formula relating the \code{response} to the \code{log_dose},
if the \link[drc:logistic]{drc::L.4} sigmoid link function is used or \code{dose} if the \link[drc:LL.4]{drc::LL.4}
log-linear sigmoid link functions used with the \code{fct} argument.}

\item{fct}{\code{Boltzmann} non-linear link function. Default is the 4 parameter
log-linear sigmoid equation, \link[drc:logistic]{drc::L.4}, where all parameters are to be
estimated. To fix a subset them, set the parameters to constant
values.}

\item{...}{additional arguments passed to \link[drc:drm]{drc::drm}}
}
\value{
An object of (S3) class \code{drc} from the \link{drc} package.

\examples
\dontrun{
BayesPharma::drm_model(
  data = tibble::tibble(
    log_dose = seq(-7, -5, length.out = 20),
    mean_response = BayesPharma::sigmoid(
      ac50 = -6, hill = -1, top = 1, bottom = 0, log_dose = log_dose),
    response = stats::rnorm(n = 20, mean = 0, sd = 0.2))
}
}
\description{
Wrapper for DRC dose response models
}
