% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{eval_init}
\alias{eval_init}
\title{Evaluate an init}
\usage{
eval_init(init, chains = NULL)
}
\arguments{
\item{init}{A valid init argument for \link[rstan:rstan]{rstan::rstan}.}

\item{chains}{\code{numeric} number of chains to generate initializations for}
}
\value{
\code{list} of \code{lists} form of model initialization.
}
\description{
How brms models can be initialized depends on the backend. The
method that all backends supports is as a list (one for each chain) of lists
(one for each variable) with numeric values. Since this requires knowing
how many chains are being run, which may not be available when the model is
being defined, and to support random initialization, the rstan backend
also supports initialization as a function returning a list of functions
(one for each parameter) returning a numeric array of length 1. Also, to
suppor the common use-case of initializing everything to zero or randomly in
the range (-2, 2) on the unconstrained scale, rstan also supports
initializing with \code{0} and \code{"random"}.

To make BayesPharma more backend agnostic, this helper function takes the
an init and the number of chains and reduces it to the list of list format.
}
