% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{eval_init_one_param}
\alias{eval_init_one_param}
\title{Helper function for \code{eval_init_one_param}}
\usage{
eval_init_one_param(param_init, param_name, sdata)
}
\arguments{
\item{param_init}{\code{array}, \code{numeric}, or \code{function} to initialize parameter}

\item{param_name}{character parameter name defined in the stan model}

\item{sdata}{\code{list} generated by Stan for the model data info}
}
\value{
\code{numeric} or \code{array}. If the parameter is of the form \verb{b_<param>}
then \link{brms} expects an array of dimension \verb{sdata$K_<param>}, otherwise
return a \code{numeric} value.
}
\description{
Helper function for \code{eval_init_one_param}
}
