% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-testthat.R
\name{expect_gg}
\alias{expect_gg}
\title{Check if an object is a ggplot2 plot}
\usage{
expect_gg(x)
}
\description{
Use in tests to check if an object is a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\note{
borrowed from bayesplot:::expect_gg
}
