% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model-cli.R
\name{fit_model_cli}
\alias{fit_model_cli}
\title{Fit model using command line arguments}
\usage{
fit_model_cli(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{\code{list} of command line arguments}
}
\description{
Provide a command line interface to fitting BayesPharma models
}
