% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-run.R
\name{fit_MuSyC_model}
\alias{fit_MuSyC_model}
\title{Combined:
     sample1 =}
\usage{
fit_MuSyC_model(
  data,
  prior = MuSyC_prior(),
  inits = MuSyC_inits(),
  formula = MuSyC_formula(),
  control = list(adapt_delta = 0.99, max_treedepth = 12),
  ...
)
}
\description{
nlp_logE0 = X_logE0 * b_logE0
                 _______   _______
                    = 1    = param
}
\details{
nlp_logE1 = X_logE1 * b_logE0
                 _______   ______
                 =sample1
}
