% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-run.R
\name{fit_MuSyC_score_by_dose_robust}
\alias{fit_MuSyC_score_by_dose_robust}
\title{Combined:
     sample1 =}
\usage{
fit_MuSyC_score_by_dose_robust(
  well_scores,
  group_vars = NULL,
  prior = MuSyC_default_prior(),
  inits = MuSyC_default_inits,
  formula = "MuSyC_combined",
  verbose = FALSE,
  iter = 8000,
  cores = 4,
  stan_model_args = list(verbose = FALSE),
  control = list(adapt_delta = 0.99, max_treedepth = 12),
  debug = FALSE,
  ...
)
}
\description{
nlp_logE0 = X_logE0 * b_logE0
                 _______   _______
                    = 1    = param
}
\details{
nlp_logE1 = X_logE1 * b_logE0
                 _______   ______
                 =sample1
}
