% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-run.R
\name{generate_MuSyC_effects}
\alias{generate_MuSyC_effects}
\title{Drug Synergy
MuSyC Drug Synergy model}
\usage{
generate_MuSyC_effects(d1, d2, E0, s1, C1, E1, s2, C2, E2, alpha, E3)
}
\arguments{
\item{d1}{Dose of drug 1}

\item{d2}{Dose of drug 2}

\item{E0}{effect with no drug treatment

# params for drug 1 by it self}

\item{s1}{drug 1 hill slope}

\item{C1}{drug 1 EC50}

\item{E1}{drug 1 maximum effect

# params for drug 2 by it self}

\item{s2}{drug 2 hill slope}

\item{C2}{drug 2 EC50}

\item{E2}{drug 2 maximum effect}

\item{beta}{synergistic efficacy
   percent increase in a drug combination’s effect
   beyond the most efficacious single drug.

   beta > 0 => synergistic efficacy
     the effect at the maximum concentration of both drugs (E3) exceeds the
     maximum effect of either drug alone (E1 or E2)

   beta < 0 => antagonistic efficacy
     at least one or both drugs are more efficacious as
     single agents than in combination}

\item{alpha1}{synergistic potency
how the effective dose of drug 1
is altered by the presence of drug 2}

\item{alpha2}{synergistic potency
   how the effective dose of drug 2
   is altered by the presence of drug 1

   alpha > 1 => synergistic potency
     the EC50 decreases because of the addition of the other drug,
     corresponding to an increase in potency

   0 <= alpha < 1 => antagonistic potency
     the EC50 of the drug increases as a result of the other drug,
     corresponding to a decrease in potency

   alpha1 == alpha2 if detailed balance}
}
\description{
Assume that the response metric decreases with more effective drugs
Let E3 be the effect at the maximum concentration of both drugs
}
\details{
Special cases:
   * dose additive model: alpha1 = alpha2 = 0
       * loewe: h1 = h2 = 1
           * CI:  E0 = 1, E1 = E2 = E3 = 0
                  the drug effect is equated with percent inhibition
   * bliss drug independence model:
       E0 = 1, E1 = E2 = E3 = 0, alpha1 = alpha2 = 1
}
