% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-run.R
\name{generate_MuSyC_effects_robust}
\alias{generate_MuSyC_effects_robust}
\title{Generate MuSyC Ed scores using a robust functional form
It should give the same results as the simple one, but
be more numerically stable}
\usage{
generate_MuSyC_effects_robust(
  logd1,
  logd2,
  logE0,
  h1,
  logC1,
  logE1,
  h2,
  logC2,
  logE2,
  logE3,
  logalpha
)
}
\description{
Generate MuSyC Ed scores using a robust functional form
It should give the same results as the simple one, but
be more numerically stable
}
