% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{get_response_variable}
\alias{get_response_variable}
\title{Get the Response Variable from a BayesPharma Model}
\usage{
get_response_variable(model)
}
\arguments{
\item{model}{\code{bpfit} object resulting from fitting a model with one of the
model functions from the \link{BayesPharma} package.}
}
\value{
\code{character} with the response variable
}
\description{
Get the Response Variable from a BayesPharma Model
}
