% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{get_treatment_units}
\alias{get_treatment_units}
\title{Get the Treatment Variable Units from a BayesPharma Model}
\usage{
get_treatment_units(model)
}
\arguments{
\item{model}{\code{bpfit} object resulting from fitting a model with one of the
model functions from the \link{BayesPharma} package.}
}
\value{
\code{character} with the treatment variable units
}
\description{
Get the Treatment Variable Units from a BayesPharma Model
}
