% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-function.R
\name{growth_richards}
\alias{growth_richards}
\title{Richards growth function}
\usage{
growth_richards(K, K0, rate, lambda, nu, time)
}
\arguments{
\item{K}{numeric, the carrying capacity}

\item{K0}{numeric, the baseline response}

\item{rate}{numeric, maximum growth rate}

\item{lambda}{numeric, time line through the maximum growth rate crosses zero}

\item{nu}{numeric}

\item{time}{numeric, time point at which to evaluate the response.}
}
\value{
numeric, response given the time and parameters


Y(time) =
  K0 + (K - K0)/(
    1 + nu * exp(
      1 + nu + rate/(K - K0) * (1 + nu)^(1 + 1/nu) *
      (lambda - time))) ^ (1/nu)
}
\description{
Functional form for the Richards growth model.
}
\examples{
\dontrun{

 # Generate Richards growth curve
 data <- data.frame(
   time = seq(0, 2, length.out = 101)) |>
     dplyr::mutate(
       response = stats::rnorm(
         n = length(time),
         mean = BayesPharma::growth_richards(
           K = 1,
           K0 = 0,
           rate = 2,
           lambda = 0.5,
           nu = 2,
           time = time),
       sd = .2))
}

}
