% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-formula.R
\name{growth_richards_formula}
\alias{growth_richards_formula}
\title{Create a Richards growth formula for the brms model}
\usage{
growth_richards_formula(
  time_variable = "time",
  time_units = "hours",
  response_variable = "response",
  response_units = NULL,
  predictors = 1,
  ...
)
}
\arguments{
\item{time_variable}{character variable representing the treatment.
(Default: 'time')}

\item{time_units}{character the units of the treatment. The default is
log base 10 of the molar concentration. (Default: "seconds")}

\item{response_variable}{character variable representing the response to
treatment. (Default: 'response')}

\item{response_units}{character the units of the response. (Default: NULL)}

\item{predictors}{Additional formula objects to specify predictors
of non-linear parameters. i.e. what perturbations/experimental
differences should be modeled separately? (Default: 1) should a
random effect be taken into consideration? i.e. cell number,
plate number, etc.}

\item{...}{additional arguments to \code{brms::brmsformula}}
}
\value{
a \code{bpformula}, which is a subclass of
    \code{brms::brmsformula} and can be passed to
    \code{BayesPharma::growth_richards_model}.
}
\description{
set-up a Richards growth model formula to define
    a non-linear model or multilevel non-linear model for
    \code{K}, \code{K0}, \code{rate}, \code{lambda}, and \code{nu} for
    use in \code{growth_richards_model} and in the BayesPharma
    package.

    This parameterization follows (Zwietering, 1990) and grofit,
}
\examples{
\dontrun{
  # Data has a string column drug_id with drug identifiers
  # Fit a separate model for each drug
  BayesPharma::growth_richards_formula(predictors = 0 + drug_id)

  # Data has a string column plate_id with plate identifiers
  # Estimate the change in response for each plate relative to a global
  # baseline.
  BayesPharma::growth_richards_formula(predictors = plate_id)

  # data has columns drug_id and plate_id
  # fit a multilevel model where the drug effect depends on the plate
  BayesPharma::growth_richards_formula(predictors = 0 + (drug_id|plate_id))
}

}
\seealso{
[brms::brmsformula()], which this function wraps.
    [BayesPharma::growth_richards_model()] into which the result of this
    function can be passed.
}
