% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-init.R
\name{growth_richards_init}
\alias{growth_richards_init}
\title{Create initialization for the Richards growth model}
\usage{
growth_richards_init(K = 1, K0 = 0, rate = 1, lambda = 0.5, nu = 2)
}
\arguments{
\item{K}{numeric or numeric returning function (default = 1).}

\item{K0}{numeric or numeric returning function (default = 0).}

\item{rate}{numeric or numeric returning function (default = 1).}

\item{lambda}{numeric units of top (default = 0.5).}

\item{nu}{numeric units of bottom (default = 2).}
}
\value{
input for brms::brm(init = ...)
}
\description{
Creating initial values for Richards growth model
    parameters that can be passed to the
    \code{richards_growth_model}.
}
\examples{
\dontrun{
init <- BayesPharma::growth_richards_init(
  A = 100,
  nu = 24)
}
}
