% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-run.R
\name{growth_richards_model}
\alias{growth_richards_model}
\title{Run Bayesian Richards Growth Model}
\usage{
growth_richards_model(
  data,
  formula = growth_richards_formula(),
  prior = growth_richards_prior(),
  init = growth_richards_init(),
  iter = 8000,
  control = list(adapt_delta = 0.99),
  stanvar_function = growth_richards_stanvar,
  expose_functions = TRUE,
  ...
)
}
\arguments{
\item{data}{\code{data.frame} of experimental data.  must contain columns
\code{response} and any predictors specified in the formula.}

\item{formula}{\code{brmsformula} object. To create a growth model formula,
(default: \code{BayesPharma::growth_formula()}.}

\item{prior}{brmspriors data.frame for A, nu, mu, and lambda.
Use one of the priors functions provided to create priors to
use here.  (default:
\code{BayesPharma::growth_richards_prior()}}

\item{init}{initial values of the parameters being modeled (default
= \code{BayesPharma::growth_richards_init()}}

\item{iter}{number of iterations the model runs. Increasing iter
can help with model convergence (default: 8000).}

\item{control}{a named list of parameters to control the sampler's
behavior.  Adding \code{max_treedepth} and giving a greater
value than 10 can improve model convergence (default:
\code{list(adapt_delta = 0.99)}).}

\item{stanvar_function}{Stan code for the model (default:
\code{BayesPharma::sigmoid_stanvar})}

\item{expose_functions}{boolean. Expose the BayesPharma functions for the
model [default: TRUE].}

\item{...}{additional arguments passed to \code{brms::brm}}
}
\value{
\code{brmsfit} object
}
\description{
For additional information on additional function arguments, reference:
  <https://paul-buerkner.github.io/brms/reference/brm.html>
  or
  <https://rdrr.io/cran/rstan/man/stan.html>
}
\examples{
\dontrun{
  BayesPharma::growth_model(
    data = data,
    formula = BayesPharma::growth_richards_formula(predictors = 0 + drug))
}
}
