% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-formula.R
\name{growth_sigmoid_formula}
\alias{growth_sigmoid_formula}
\title{Create a formula for the sigmoid growth model}
\usage{
growth_sigmoid_formula(
  treatment_variable = "time",
  treatment_units = "hours",
  response_variable = "response",
  response_units = NULL,
  predictors = 1,
  ...
)
}
\arguments{
\item{treatment_variable}{\code{character} variable representing time as a
treatment}

\item{treatment_units}{\code{character} the units of the time variable}

\item{response_variable}{\code{character} variable representing the response to
treatment}

\item{response_units}{\code{character} the units of the response}

\item{predictors}{Additional formula objects to specify predictors
of non-linear parameters. i.e. what perturbations/experimental
differences should be modeled separately? (Default: 1) should a
random effect be taken into consideration? i.e. cell number,
plate number, etc.}

\item{...}{additional arguments to \code{\link[brms:brmsformula]{brms::brmsformula()}}}
}
\value{
a \code{bpformula}, which is a subclass of
\link[brms:brmsformula]{brms::brmsformula} and can be passed to
\code{\link[=growth_sigmoid_model]{growth_sigmoid_model()}}.
}
\description{
set-up a sigmoid growth model formula to for use in
\code{\link[=growth_sigmoid_model]{growth_sigmoid_model()}}. The functional form is

\if{html}{\out{<div class="sourceCode">}}\preformatted{response ~ growth_sigmoid(K, K0, rate, lambda, time)
}\if{html}{\out{</div>}}

The parameterization follows (Zwietering, 1990) and \link{grofit}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{K      = **carrying capacity**, `K = response(time = Inf)`. The [grofit]
         package calls this parameter `A`. `K` has the same units as the
         `response`.
K0     = **initial population size** `K0 = response(time = 0)`. The
         [grofit] package assumes `K0=0`. `K0` has the same units as the
         `response`.
rate   = **maximum growth rate** `rate = max[d(response)/d(time)]`. The
         [grofit] package calls this `mu`. `rate` has the units of
         `response/time`
lambda = **duration of the lag-phase** the time point at which the
         tangent through the growth curve when it achieves the maximum
         growth rate crosses the initial population size `K0`. (see
         Figure 2 in (Kahm et al., 2010)).
}\if{html}{\out{</div>}}

See the vignettes(topic = "derive_growth_model", package = "BayesPharma")
}
\examples{
\dontrun{
  # Data has a string column drug_id with drug identifiers
  # Fit a separate model for each drug
  BayesPharma::growth_sigmoid_formula(predictors = 0 + drug_id)

  # Data has a string column plate_id with plate identifiers
  # Estimate the change in response for each plate relative to a global
  # baseline.
  BayesPharma::growth_sigmoid_formula(predictors = plate_id)

  # data has columns drug_id and plate_id
  # fit a multilevel model where the drug effect depends on the plate
  BayesPharma::growth_sigmoid_formula(predictors = 0 + (drug_id|plate_id))
}

}
\references{
Zwietering M. H., Jongenburger I., Rombouts F. M., van 't Riet K., (1990)
Modeling of the Bacterial Growth Curve. Appl. Environ. Microbiol., 56(6),
1875-1881 https://doi.org/10.1128/aem.56.6.1875-1881.1990

Kahm, M., Hasenbrink, G., Lichtenberg-Fraté, H., Ludwig, J., & Kschischo, M.
(2010). grofit: Fitting Biological Growth Curves with R. J. Stat. Softw.,
33(7), 1–21. https://doi.org/10.18637/jss.v033.i07
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{[brms::brmsformula()], which this function wraps.
[growth_sigmoid_model]()] into which the result of this
function can be passed. Related to [grofit::logistic]
}\if{html}{\out{</div>}}
}
