% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-init.R
\name{growth_sigmoid_init}
\alias{growth_sigmoid_init}
\title{Create Initialization for the Sigmoid Growth Model}
\usage{
growth_sigmoid_init(
  K = function() runif(n = 1, min = 0.8, max = 1.2),
  K0 = function() runif(n = 1, min = -0.2, max = 0.2),
  rate = function() runif(n = 1, min = 0.8, max = 1.2),
  lambda = function() runif(n = 1, min = 0.3, max = 0.7),
  ...
)
}
\arguments{
\item{K}{\code{numeric} or \code{numeric} returning \code{function}}

\item{K0}{\code{numeric} or \code{numeric} returning \code{function}}

\item{rate}{\code{numeric} or \code{numeric} returning \code{function}}

\item{lambda}{\code{numeric} or \code{numeric} returning \code{function}
units of \code{time}}

\item{...}{additional parameter initialization. Each named argument should be
a function that returns a \code{numeric} or \code{array} depending on the dimension
of the parameter, see \code{\link[=rstan_default_init]{rstan_default_init()}} to use the \pkg{rstan} default
init.}
}
\value{
input for \verb{[brms::brm](init = ...)}
}
\description{
Create initial values for sigmoid growth model parameters that
can be passed to the \link{growth_sigmoid_model}.
}
\examples{
\dontrun{
init <- BayesPharma::growth_sigmoid_init(
  A = 100)
}
}
\seealso{
\link{growth_sigmoid_formula}, \link{growth_sigmoid_prior}, and
\link{growth_sigmoid_model}
}
