% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-prior.R
\name{growth_sigmoid_prior}
\alias{growth_sigmoid_prior}
\title{Create a Prior for the Sigmoid Growth Model}
\usage{
growth_sigmoid_prior(
  K = brms::prior(prior = normal(1, 0.5), nlpar = "K"),
  K0 = brms::prior(prior = normal(0, 0.5), nlpar = "K0"),
  rate = brms::prior(prior = normal(2, 0.5), nlpar = "rate", lb = 0),
  lambda = brms::prior(prior = normal(0.5, 0.5), nlpar = "lambda"),
  ...
)
}
\arguments{
\item{K}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
\code{K} parameter. (Default: \code{normal(1, 0.5)})}

\item{K0}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
\code{K0} parameter. (Default: \code{normal(0, 0.5)})}

\item{rate}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
\code{rate} parameter. (Default: \code{normal(1, 1)} with and a lower bound
of \code{0}).}

\item{lambda}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
\code{lambda} parameter. (Default: \code{normal(0.5, 0.5)}).}

\item{...}{additional \code{\link[brms:set_prior]{brms::brmsprior()}} objects.}
}
\value{
\code{\link[brms:set_prior]{brms::brmsprior()}} \code{data.frame}
}
\description{
Creates a for the priors for the sigmoid growth model. If the
function arguments \code{K}, \code{K0}, \code{rate}, \code{lambda}, or
\code{NULL}, normal distribution priors will be set.

If you would like to set a parameter as a constant, enter a numeric  value
for the function argument.

For other distribution options, reference \link[brms:set_prior]{brms::prior} and \href{https://github.com/stan-dev/stan/wiki/Prior-Choice-Recommendations}{Prior Choice Recommendations }
}
\examples{
\dontrun{
priors <- growth_sigmoid_prior(
  K = brms::prior(prior = normal(100, 20), nlpar = "K"),
  lambda = 1)
}
}
