% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-stanvars.R
\docType{data}
\name{growth_sigmoid_stanvar}
\alias{growth_sigmoid_stanvar}
\title{Stan code for the growth Richards function.}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
growth_sigmoid_stanvar
}
\description{
Stan code for the growth Richards function.
}
\note{
stanvar script of a sigmoid growth model response equation
}
\keyword{datasets}
