% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kor_antag}
\alias{kor_antag}
\title{Dose Response Measurements of Four Different Kappa Opioid Receptor
Antagonists}
\format{
A \code{data.frame} that includes the dose-response measurements
}
\source{
\link{https://github.com/maomlab/BayesPharma}
}
\usage{
kor_antag
}
\description{
A data set containing the drugs used, cell id, the log dose that the response
was measured at, and the reference, raw, and normalized measurement
}
\keyword{datasets}
