% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-michaelis_menten-generate.R
\name{michaelis_menten}
\alias{michaelis_menten}
\title{Generate Data from the Michaelis Menten Enzyme Kinetics}
\usage{
michaelis_menten(time, kcat, kM, ET, ST, ...)
}
\arguments{
\item{time}{\code{numeric} vector. Increasing time points}

\item{kcat}{\code{numeric} value catalytic rate constant}

\item{kM}{\code{numeric} value Michaelis rate constant}

\item{ET}{\code{numeric} value total enzyme concentration}

\item{ST}{\code{numeric} value total substrate concentration}

\item{...}{additional arguments to \code{\link[deSolve:ode]{deSolve::ode()}}}
}
\value{
run the Michaelis Menten ordinary differential equation forwards
starting with initial product concentration of \code{0} and specified \code{kcat} and
\code{kM} parameters for the specified time steps.
}
\description{
Generate data from the Michaelis Menten enzyme kinetics ordinary differential
equation described in (Choi, et al., 2017, DOI: 10.1038/s41598-017-17072-z).
Consider the kinetic rate equation
\preformatted{
                  kf
                 --->    kcat
          E + S  <---  C --->  E + P
                  kb}
}
\details{
where the free enzyme (E) reversibly binds to the substrate (S) to
form a complex (C) with forward and backward rate constants of kf
and kb, which is irreversibly catalyzed into the product (P), with
rate constant of kcat, releasing the enzyme to catalyze additional
substrate. The total enzyme concentration is defined to be the
\code{ET := E + C}. The total substrate and product concentration
is defined to be \code{ST := S + C + P}. The Michaelis constant is
the defined to be the \code{kM := (kb + kcat) / kf}. The kcat rate
constant determines the maximum turn over at saturating substrate
concentrations, \code{Vmax := kcat * ET}. The rate constants \code{kcat}
and \code{kM} can be estimated by monitoring the product accumulation
over time (enzyme progress curves), by varying the enzyme and
substrate concentrations.

From (Choi, et al, 2017, equation 2, the Michaelis Menten differential
equation is defined by
\preformatted{
  Observed data:
     M     = number of measurements        # number of measurements
     t[M]  = time                          # measured in seconds
     Pt[M] = product                       # product produced at time t
     ST    = substrate total concentration # specified for each experiment
     ET    = enzyme total concentration    # specified for each experiment

  Model parameters:
    kcat    # catalytic constant (min^-1)
    kM      # Michaelis constant ()

  ODE formulation:
    dPdt = kcat * (
             ET + kM + ST - Pt +
             -sqrt((ET + kM + ST - Pt)^2 - 2 * ET * (ST - Pt))) / 2

  initial condition:
     P := 0}

In (Choi, et al. 2017) they prove, that the tQ model is valid when
\preformatted{
    K/(2*ST) * (ET+kM+ST) / sqrt((ET+kM+ST+P)^2 - 4*ET(ST-P)) << 1,}

where K = kb/kf is the dissociation constant.
}
\references{
Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation:
Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7,
17018 (2017). https://doi.org/10.1038/s41598-017-17072-z
}
\seealso{
\link{michaelis_menten_model}
}
