% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-michaelis_menten-formula.R
\name{michaelis_menten_formula}
\alias{michaelis_menten_formula}
\title{Define a Formula for the Michaelis Menten Enzyme Kinetics Model}
\usage{
michaelis_menten_formula(
  series_index_variable = "series_index",
  treatment_variable = "time",
  treatment_units = "seconds",
  ET_variable = "ET",
  ET_units = "mg/ml",
  ST_variable = "ST",
  ST_units = "mg/ml",
  response_variable = "P",
  response_units = "mg/ml",
  predictors = 1,
  ...
)
}
\arguments{
\item{series_index_variable}{\code{character} variable indexing which
measurements are part of a common time series}

\item{treatment_variable}{\code{character} variable representing time as the
treatment}

\item{treatment_units}{\code{character} the units of the treatment variable}

\item{ET_variable}{\code{character} variable representing the total enzyme
concentration}

\item{ET_units}{\code{character} variable representing the units of the
enzyme concentration}

\item{ST_variable}{\code{character} variable representing the total substrate
concentration}

\item{ST_units}{\code{character} variable representing the units of the
substrate concentration}

\item{response_variable}{\code{character} variable representing the \code{product}}

\item{response_units}{\code{character} the units of the \code{product}}

\item{predictors}{\code{expression} of predictors for \code{kcat} and \code{kM} parameters}

\item{...}{additional arguments passed to \code{\link[brms:brmsformula]{brms::brmsformula()}}}
}
\value{
a \code{bpformula}, which is a subclass of \link[brms:brmsformula]{brms::brmsformula} and can be
passed to \code{\link[=michaelis_menten_model]{michaelis_menten_model()}}.
}
\description{
set-up a Michaelis Menten enzyme kinetics model formula to
define a an ordinary differential equation model that with \code{kcat}, \code{kM}
model parameters for use in the \link{michaelis_menten_model}.
}
\references{
Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation:
Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7,
17018 (2017). https://doi.org/10.1038/s41598-017-17072-z
}
\seealso{
\link{michaelis_menten_model}, \link{michaelis_menten_prior},
\link{michaelis_menten_init}, \link{michaelis_menten_stanvar} and
\link[brms:brmsformula]{brms::brmsformula}, which this function wraps.
}
