% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-michaelis_menten-stanvars.R
\docType{data}
\name{michaelis_menten_genquant_stanvar}
\alias{michaelis_menten_genquant_stanvar}
\title{Stan Code for the Michaelis Menten Model Generated Quantities}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
michaelis_menten_genquant_stanvar
}
\description{
If only the substrate concentration is varied, it is not generally possible
to fit both \code{kcat} and \code{kM}. However, it is possible to fit the ratio
\code{kcat/kM}. Including this \link{stan} code will generate the samples for
\code{kcat/kM}.
}
\references{
Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation:
Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7,
17018 (2017). https://doi.org/10.1038/s41598-017-17072-z
}
\seealso{
\link{michaelis_menten_model}
}
\keyword{datasets}
