% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-michaelis_menten-prior.R
\name{michaelis_menten_prior}
\alias{michaelis_menten_prior}
\title{Create a Prior for the Michaeli Menten Enzyme Kinetics Model}
\usage{
michaelis_menten_prior(
  kcat = brms::prior_string(prior = "gamma(4, 1)", lb = 0, nlpar = "kcat"),
  kM = brms::prior_string(prior = "gamma(4, 1)", lb = 0, nlpar = "kM")
)
}
\arguments{
\item{kcat}{prior for \code{kcat} parameter. Given a numeric value, it will be
used as a constant and not estimated.}

\item{kM}{prior for \code{kM} parameter. Given a numeric value, it will be used as
a constant and not estimated.}
}
\description{
For both the \code{kcat} and \code{kM} parameters, the default priors are
\code{gamma(4, 1)}. We use the gamma distribution because it is naturally lower
bounded by \code{0}. The first parameter is the shape, \code{alpha=4}, and the second
is the rate, \code{beta=1}. The mean of gamma distributions is \code{alpha/beta} and
the variance is \code{alpha/beta^2}.
}
\seealso{
\link{michaelis_menten_model}, \link{michaelis_menten_formula},
\link{michaelis_menten_init}, and \link{michaelis_menten_stanvar}
}
