% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-michaelis_menten-stanvars.R
\name{michaelis_menten_stanvar}
\alias{michaelis_menten_stanvar}
\title{Stan Code for the Michaelis Menten Kinetics Model}
\usage{
michaelis_menten_stanvar()
}
\description{
The Michaelis Menten model is an ordinary differential equation
model for the change in product as a function of the total enzyme
concentration (ET), total substrate concentration (ST), the Michaelis
constant (kM) and the catalytic constant (kcat). To implement the Michaelis
Menten model in \link[rstan:stan]{rstan::stan}, the function \code{michaelis_menten_ode} is defined
and then passed to \code{michaelis_menten_single} to integrate it using the \emph{stiff
backward differentiation formula (BDF) method}. To fit multiple time series
in one model, the \code{michaelis_menten_multiple} can be used. Note that to
handle fitting time-series with different numbers of observations, an
additional \code{series_index} argument is used. Note that observations in the
same time-series should be in sequential order in the supplied data.
}
\examples{
\dontrun{
brms::brm(
  data = ...,
  formula = brms::brmsformula(
    P ~ michaelis_menten_multiple(
      series_index, time, kcat, kM, ET, ST),
    kcat + kM ~ 1,
    nl = TRUE,
    loop = FALSE),
  prior = ...,
  init =  ...,
  stanvars = BayesPharma::michaelis_menten_stanvar())
}

}
\references{
Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation:
Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7,
17018 (2017). https://doi.org/10.1038/s41598-017-17072-z
}
\seealso{
\link{michaelis_menten_model},
}
