% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-formula.R
\name{MuSyC_formula}
\alias{MuSyC_formula}
\title{Create a formula for the MuSyC synergy model}
\usage{
MuSyC_formula(predictors = 1, ...)
}
\arguments{
\item{predictors}{Additional formula objects to specify predictors of
non-linear parameters. i.e. what perturbations/experimental differences
should be modeled separately? (Default: 1) should a random effect be taken
into consideration? i.e. cell number, plate number, etc.}
}
\value{
brmsformula
}
\description{
setup a defaulMuSyC synergy model formula to predict
  the `E0`, `C1`, `E1`, `s1`, `C2`, `E2`, `s2`, `log10alpha`, and `E3alpha`
  parameters.
}
\examples{
\dontrun{
  Consider observations made using 4 different drugs and the column header
  containing the labels for the 4 different drugs is `predictors`.

    `sigmoid_formula(predictors = 0 + predictors)`

  Consider that the cell_ID was recorded and the noise from the cell_ID is to
  be accounted for. 
  
   `sigmoid_formula(predictors = 0 + predictors + (1|cell_ID))`
}

}
