% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-stanvars.R
\docType{data}
\name{MuSyC_genquant_stanvar}
\alias{MuSyC_genquant_stanvar}
\title{Stan Code for the MuSyC Generated Quantities}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
MuSyC_genquant_stanvar
}
\description{
Stan Code for the MuSyC Generated Quantities
}
\keyword{datasets}
