% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-utils.R
\name{MuSyC_hi_to_si}
\alias{MuSyC_hi_to_si}
\title{Convert Exponent to Slope Parameterization for Drug i in MuSyC Model}
\usage{
MuSyC_hi_to_si(hi, Ci, E0, Ei)
}
\arguments{
\item{hi}{\code{numeric} value of the exponent in the MuSyC equation for drug
i}

\item{Ci}{\code{numeric} value of the AC50 of drug i}

\item{E0}{\code{numeric} value of the response with no treatments}

\item{Ei}{\code{numeric} value of the response of infinite drug i and no
other treatments}
}
\value{
si slope of drug i at it's AC50 and doses of all other drugs are
zero
}
\description{
This can be used for setting priors and interpreting parameter
estimates see \code{\link[=MuSyC_si_to_hi]{MuSyC_si_to_hi()}} for details
}
\seealso{
\link{MuSyC_si_to_hi}
}
