% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-utils.R
\name{MuSyC_si_to_hi}
\alias{MuSyC_si_to_hi}
\title{Convert From Slope to Exponent Parameterization for Drug i in the MuSyC Model}
\usage{
MuSyC_si_to_hi(si, Ci, E0, Ei)
}
\arguments{
\item{si}{\code{numeric} value of slope of drug i at its AC50 and doses of
all other drugs are zero}

\item{Ci}{\code{numeric} value of the AC50 of drug i}

\item{E0}{\code{numeric} value of the response with no treatments}

\item{Ei}{\code{numeric} value of the response of infinite drug i and no
other treatments}
}
\value{
hi \code{numeric} value of the exponent in the MuSyC equation for
drug i
}
\description{
This can be used for setting priors and interpreting parameter
estimates
}
\details{
Claim: When \code{d1=0} and \code{d2=C2} then the gradient of the response
with respect to \code{d2} is the \code{s2}, symbolically \code{d(Ed)/d(d2) = s2} where
\code{s2 = h2 * (E0 + E2) / (4 * C2)} then

\if{html}{\out{<div class="sourceCode">}}\preformatted{d(Ed)/d(d2)
  = d/d(d2)
    (C1^h1 * C2^h2 * E0 + C1^h1 * d2^h2 * E2) /
    (C1^h1 * C2^h2      + C1^h1 * d2^h2)
}\if{html}{\out{</div>}}

Cancel the \code{C1^h1} terms:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  =  d/d(d2)
     (C2^h2 * E0 + d2^h2 * E2) /
     (C2^h2      + d2^h2)
}\if{html}{\out{</div>}}

Distribute the derivative across the terms in the numerator

\if{html}{\out{<div class="sourceCode">}}\preformatted{  =  E0 * C2^h2 * (d/d(d2) 1     / (C2^h2 + d2^h2)) +
     E2         * (d/d(d2) d2^h2 / (C2^h2 + d2^h2))

  =  E0 * C2^h2 * (h2 * d2^(h2-1) / (C2^h2 + d2^h2)^2) +
     E2 * (C2^h2 * h2 * d2^(h2-1) / (C2^h2 + d2^h2)^2)

  =  (E0 + E2) * C2^h2 * h2 * d2^(h2-1)/(C2^h2 + d2^h2)^2
}\if{html}{\out{</div>}}

Evaluate at \code{d2 = C2}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  =  (E0 + E2) * h2 * C2^(2*h2-1) / (4*C2^(2*h2)))
  =  h2 * (E0 + E2) / (4 * C2)
}\if{html}{\out{</div>}}
}
\seealso{
\link{MuSyC_hi_to_si}
}
