% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-stanvars.R
\name{MuSyC_stanvar}
\alias{MuSyC_stanvar}
\title{Stan Code for the MuSyC Function.}
\usage{
MuSyC_stanvar()
}
\description{
Stan Code for the MuSyC Function.
}
\note{
stanvar script of a MuSyC equation.
}
