% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_model_draws.R
\name{n_sample_draws}
\alias{n_sample_draws}
\title{Dataframe of posterior samples of modeled parameters from brmsfit model.}
\usage{
n_sample_draws(model = <brmsfit_model_name>,
               n = 100)
}
\arguments{
\item{model}{brmsfit model.}

\item{n}{numeric units of samples from the posterior distribution (default n = 100).}
}
\value{
data.frame.
}
\description{
Given a brms model, extract a sample of values from posterior distribution.
}
\examples{
Consider a brmsfit model named "activator_dr_model" and the number of samples wanted is 50.
n_sample_draws(model = activator_dr_model, n = 50)

}
