% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dr_brm_model.R
\name{no_predictors_dr_formula}
\alias{no_predictors_dr_formula}
\title{Create a formula for a singular perturbation/experimental variable for the
brms model}
\usage{
no_predictors_dr_formula(...)
}
\value{
brmsformula
}
\description{
additional arguments for brmsformula can be added.
Reference this website for more information on additional arguments:
https://paul-buerkner.github.io/brms/reference/brmsformula.html
}
