% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_model_draws.R
\name{plot_draws_data}
\alias{plot_draws_data}
\title{Create a plot of the predicted responses from the posterior distribution}
\usage{
plot_draws_data(
  model,
  n = 100,
  lower = -12,
  upper = -3,
  predictor_name,
  data,
  measurement,
  draws,
  pred_response,
  mean_draws,
  title = NULL,
  xlabel = "Log Dose",
  ylabel = "Response"
)
}
\arguments{
\item{model}{brmsfit model.}

\item{n}{number of samples from the posterior distribution (default n = 100).}

\item{lower}{number for the lowest log_dose value to be used for calculating Response (default = -12).}

\item{upper}{number for the highest log_dose value to be used for calculating Response (default = -3).}

\item{data}{is the tibble or data.frame used for the brmsfit.}

\item{measurement}{is the column in 'data' containing response values.}

\item{draws}{is the tibble::tibble returned from the posterior_response_draws' function.}

\item{pred_response}{is the 'Response' column in the tibble::tibble returned from the 'posterior_response_draws' function.}

\item{mean_draws}{is the tibble::tibble returned from the 'posterior_mean' function.}

\item{title}{string for the plot title. (default = NULL)}

\item{xlabel}{string for the x-axis label. (default = NULL)}

\item{ylabel}{string for the y-axis label. (default = NULL)}

\item{predictors_name}{string. If only one perturbation is being tested, the
name of the predictor being analyzed is required to avoid "NA" graph subtitle (default = NULL).}
}
\value{
a ggplot2::ggplot object.
}
\description{
This function holds the 3 functions required to plot the trajectory curves.
}
