% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drc.R
\name{plot_drc}
\alias{plot_drc}
\title{Create plot of drc model}
\usage{
plot_drc(
  model,
  newdata = NULL,
  predict_args = list(interval = "prediction"),
  aes_mapping = ggplot2::aes(x = log_dose, y = response),
  title = NULL
)
}
\arguments{
\item{model}{drc object. drc model generated with drc::drm(...)}

\item{newdata}{data.frame where the predictions should be made. Default: use
the range of log_dose values in the model object.}

\item{predict_args}{list of arguments passed to predict(model, ...)
including interval will generate ribbon of the given type, see
`drc:::predict.drc` for more details, Default:
`list(interval = "prediction")`}

\item{aes_mapping}{ggplot2::aes(...) mapping columns of the model data
and new data to the plot. Default:
`ggplot2::aes(x = log_dose, y = response)`}

\item{title}{character, plot title. Default: NULL}
}
\description{
Create plot of drc model
}
