% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-density.R
\name{plot_posterior_density}
\alias{plot_posterior_density}
\title{Create a plot of the posterior density distributions of modeled parameters
from brmsfit model}
\usage{
plot_posterior_density(
  model,
  predictors_col_name = "_Intercept",
  half_max_label = "ac50",
  l_ci = 0.025,
  u_ci = 0.975,
  title_label = "Posterior Density Plots w/ Mean & 95\% CI"
)
}
\arguments{
\item{model}{\code{bpfit} model.}

\item{predictors_col_name}{\code{character} of the predictors column in the input
\code{data.frame}. Predictors are the perturbations tested during the experiment
(i.e. Drug, Temperature, etc.).}

\item{half_max_label}{\code{character} of the label for the half maximal that
fits the type of experiment that was done (i.e. ec50, ic50, ed50, id50,
ld50, etc.).}

\item{l_ci}{\code{numeric} unit of the lower confidence interval}

\item{u_ci}{\code{numeric} unit of the upper confidence interval}

\item{title_label}{string of the plot title.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
plots the \code{\link[=bpfit]{bpfit()}} model density distribution of each
parameter and the mean, lower confidence interval, and upper confidence
interval.
}
\examples{
\dontrun{
  plot_posterior_density(
    model = my_sigmoid_model,
    predictors_col_name = "predictors",
    half_max_label = "ic50",
    l_ci = 0.025,
    u_ci = 0.975,
    title_label = "Posterior Density Plots with Mean and 95\% CI")
}
}
