% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_posterior_check.R
\name{plot_pp_check}
\alias{plot_pp_check}
\title{Posterior Predictive Check}
\usage{
plot_pp_check(model, plot_type = "dens_overlay", n = 50, ...)
}
\arguments{
\item{model}{brmsfit.}

\item{plot_type}{string. Name of available plot types.}

\item{n}{numeric value. Number of posterior draws to be used.}
}
\value{
ggplot object.
}
\description{
pp_check from the brms package has different plot types to analyze the model
fit by comparing the observed data with generated data from the model.
}
\details{
Here is the documentation for pp_check: http://paul-buerkner.github.io/brms/reference/pp_check.brmsfit.html

Here is the documentation for pp_check plot types under
'PPC plotting functions': https://mc-stan.org/bayesplot/reference/PPC-overview.html
The plot function can be entered as a string for argument 'plot_type' in the
function
}
