% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prior_posterior_densities.R
\name{plot_prior_posterior_densities}
\alias{plot_prior_posterior_densities}
\title{Plot Prior and Posterior Marginal Distributions}
\usage{
plot_prior_posterior_densities(
  model,
  title_label = "Prior Posterior Density",
  exclude_variables = c("__$", "lprior"),
  ...
)
}
\arguments{
\item{model}{\code{bpfit} object resulting from fitting a BayesPharma model}

\item{title_label}{string of the plot title.}

\item{exclude_variables}{\code{string} vector of variables to exclude}

\item{...}{further arguments passed to \code{\link[brms:brm]{brms::brm()}} to sample from the
prior}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generate a plot that shows for each model variable the
distribution of the marginal posterior uncertainty overlaid on the
distribution of the marginal prior uncertainty.
}
\examples{
\dontrun{
  plot_prior_posterior_densities(
    model = my_sigmoid_model,
    title_label = "Prior Posterior Density Plots")
}


}
