% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-function.R
\name{plot_sigmoid_functional_form}
\alias{plot_sigmoid_functional_form}
\title{Plot the Sigmoid Functional Form with Labeled Parameters}
\usage{
plot_sigmoid_functional_form(
  ac50,
  hill,
  top,
  bottom,
  log_dose,
  treatment_units,
  response_units
)
}
\arguments{
\item{ac50}{\code{numeric} value for the \code{ac50} parameter, which is either \code{ec50}
with the \code{\link[=sigmoid_agonist_formula]{sigmoid_agonist_formula()}} or \code{ic50} for the
\code{\link[=sigmoid_antagonist_formula]{sigmoid_antagonist_formula()}}}

\item{hill}{\code{numeric} value for the \code{hill} parameter}

\item{top}{\code{numeric} value for the \code{top} parameter}

\item{bottom}{\code{numeric} value for the \code{bottom} parameter}

\item{log_dose}{\code{numeric} vector of treatment values}

\item{treatment_units}{\code{character} units for the log dose}

\item{response_units}{\code{character} units for the response}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Generates a plot of the \code{\link[=sigmoid]{sigmoid()}} functional form with the values
of the parameters \code{ac50}, \code{hill}, \code{top}, and \code{bottom}
labeled.
}
\examples{
\dontrun{
  plot_sigmoid_functional_form(
    ac50 = -6,
    hill = -1,
    top = 100,
    bottom = 0,
    treatment_label = "Log[Molar]",
    response_label = "\% Baseline")}

}
\seealso{
\code{\link[=sigmoid]{sigmoid()}}
}
