% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-synergy-checkerboard.R
\name{plot_synergy_checkerboard}
\alias{plot_synergy_checkerboard}
\title{Plot Synergy Checkerboard}
\usage{
plot_synergy_checkerboard(
  data,
  treatment_1_label = "Treatment 1",
  treatment_2_label = "Treatment 2",
  treatment_1_units = NULL,
  treatment_2_units = NULL,
  plot_zero_dose = TRUE,
  contour_color = "gold"
)
}
\arguments{
\item{data}{\code{data.frame} with columns \code{dose1}, \code{dose2}, and \code{response}}

\item{treatment_1_label}{\code{character} used to make default title and axis
labels}

\item{treatment_2_label}{\code{character} used to make default title and axis
labels}

\item{treatment_1_units}{\code{character} used to make default axis labels}

\item{treatment_2_units}{\code{character} used to make default axis labels}

\item{plot_zero_dose}{\code{logical} plotted on the log scale, zero doses would be
at \code{-Inf}, so to show them on the plot, add them as with a slight
separation on the axis.}

\item{contour_color}{\code{character} the color of the contour lines}
}
\value{
\link[ggplot2:ggplot]{ggplot2::ggplot} plot with light-blue to dark-blue tiles the dose
response. Individual plot elements can be over-written and the
plot can be saved with \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
}
\description{
Plot Synergy Checkerboard
}
