% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_model_draws.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Create a plot of the predicted responses from the posterior distribution}
\usage{
plot_trajectories(
  data,
  measurement,
  predictors_col_name = "_Intercept",
  draws,
  pred_response,
  mean_draws,
  title = NULL,
  xlabel = "Log Dose",
  ylabel = "Response"
)
}
\arguments{
\item{data}{is the tibble or dataframe used for the brmsfit.}

\item{measurement}{is the column in 'data' containing response values.}

\item{predictors_col_name}{string expression for predictors column in the
input data.frame (default = "_Intercept").}

\item{draws}{is the tibble::tibble returned from the
'posterior_response_draws' function.}

\item{pred_response}{is the 'Response' column in the tibble::tibble returned
from the 'posterior_response_draws' function.}

\item{mean_draws}{is the tibble::tibble returned from the 'posterior_mean'
function.}

\item{title}{string for the plot title. (default = NULL)}

\item{xlabel}{string for the x-axis label. (default = NULL)}

\item{ylabel}{string for the y-axis label. (default = NULL)}
}
\value{
a ggplot2::ggplot object.
}
\description{
Create a plot of the predicted responses from the posterior distribution
}
