% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_density_distribution_plots.R
\name{posterior_densities}
\alias{posterior_densities}
\title{Create a plot of the posterior density distributions of modeled parameters from brmsfit model}
\usage{
posterior_densities(model = <model>,
                      predictors_col_name = <predictors_col_name>,
                      half_max_label = <half_max_label>,
                      l_ci = <l_ci>,
                      u_ci = <u_ci>,
                      title_label = <title_label>)
}
\arguments{
\item{model}{brmsfit model.}

\item{predictors_col_name}{string expression for predictors column in the input data.frame (default = "_Intercept").
Predictors are the perturbations tested during the experiment (i.e. Drug, Temperature, etc.).}

\item{half_max_label}{string of the label for the half maximal that fits the type of
experiment that was done (i.e. ec50, ic50, ed50, id50, ld50, etc.).}

\item{l_ci}{numeric unit of the lower confidence interval (default = 0.025)}

\item{u_ci}{numeric unit of the upper confidence interval (default = 0.975)}

\item{title_label}{string of the plot title. (default = "Posterior Density Plots with Mean and 95% CI")}
}
\value{
ggplot2::ggplot object.
}
\description{
plots the brmsfit model density distribution of each parameter
  and the mean, lower confidence interval, and upper confidence interval.
}
\examples{
  posterior_densities(model = my_dr_model,
                      predictors_col_name = "predictors",
                      half_max_label = "ic50",
                      l_ci = 0.025,
                      u_ci = 0.975,
                      title_label = "Posterior Density Plots with Mean and 95\% CI")

}
