% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-draws.R
\name{posterior_draws_plot}
\alias{posterior_draws_plot}
\title{Plot of Posterior Model Fit Draws}
\usage{
posterior_draws_plot(
  model,
  newdata = NULL,
  treatment_variable = NULL,
  treatment_units = NULL,
  response_variable = NULL,
  response_units = NULL,
  title = "Dose-Response Posterior Draws",
  n = 50,
  point_size = 0.75,
  jitter_height = 0,
  jitter_width = 0
)
}
\arguments{
\item{model}{bpfit object resulting from fitting a BayesPharma model}

\item{newdata}{data.frame of newdata to use for
predictions. Default data.frame with each predictor and
treatment variable.}

\item{treatment_variable}{string or NULL. If NULL the treatment variable
name will be looked up from the model formula. The treatment variable
the model$data or newdata if supplied must have a column corresponding to
the treatment variable (Default: NULL).}

\item{treatment_units}{string or NULL. If NULL the treatment units will
be looked up in the model formula. The treatment units will be used to
label the X-axis of the plot. (Default: NULL)}

\item{response_variable}{string or NULL. If NULL the response variable
name will be looked up from the model formula. The response variable
the model$data or newdata if supplied must have a column corresponding to
the response variable (Default: NULL).}

\item{response_units}{string or NULL. If NULL the response units will
be looked up in the model formula. The response units will be used to
label the X-axis of the plot. (Default: NULL)}

\item{title}{character name for the plot (default =
"Dose-Response Posterior Draws").}

\item{n}{numeric value of the number of draws to be observed
(default = 50).}

\item{point_size}{numeric. \code{geom_jitter} point size (default =
0.75).}

\item{jitter_height}{numeric. the height distance between
overlapping points (default = 0).}

\item{jitter_width}{numeric. the width distance between overlapping
points (default = 0).}
}
\value{
ggplot2::ggplot object.
}
\description{
A plot of a sample of model fit draws from the
    posterior distribution from the expected mean and median
    quantile intervals.
}
\examples{
\dontrun{
  # Consider a model named my_model and data named my_data with a column
  # named predictors containing multiple different perturbations.
  posterior_draws_plot(
    model = my_model,
    newdata = my_data,
    predictors_col_name = "predictors",
    lower = -12,
    upper = -3,
    n = 50,
    facet_var = predictors,
    jitter_width = 0.10,
    title = "Dose-Response Posterior Draws",
    xlab = "Log[Molar]",
    ylab = "Response")
}

}
