% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_model_draws.R
\name{posterior_mean}
\alias{posterior_mean}
\title{Tibble of predictors, mean ec50, mean hill, mean top, mean bottom, and
mean Response of the posterior distribution.}
\usage{
posterior_mean(model, n = 100, lower = -12, upper = -3, predictor_name = NULL)
}
\arguments{
\item{model}{brmsfit.}

\item{n}{number of samples from the posterior distribution. (default n = 100)}

\item{lower}{number for the lowest log_dose value to be used for calculating
Response. (default = -12)}

\item{upper}{number for the highest log_dose value to be used for calculating
Response. (default = -3)}

\item{predictors_name}{string. If only one perturbation is being tested, the
name of the predictor being analyzed is required to avoid "NA" graph subtitle.
(default = NULL)}
}
\value{
tibble::tibble required for the 'plot_trajectories' function.
}
\description{
Tibble of predictors, mean ec50, mean hill, mean top, mean bottom, and
mean Response of the posterior distribution.
}
