% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dr_brm_model.R
\name{predictors_dr_formula}
\alias{predictors_dr_formula}
\title{Create a formula for a multiple different perturbations/experimental variables
for the brms model}
\usage{
predictors_dr_formula(predictors = 0 + predictors, ...)
}
\arguments{
\item{predictors}{expression of the grouping variables for the parameters.
i.e. what perturbations/experimental differences should be modeled separately?}
}
\description{
Create a formula for a multiple different perturbations/experimental variables
for the brms model
}
