% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{prepare_init}
\alias{prepare_init}
\title{Prepare an init for a brms model}
\usage{
prepare_init(init)
}
\arguments{
\item{init}{function returning an numeric array of length one or a
numeric value.}
}
\value{
function returning a numeric array of length 1.
}
\description{
Prepare an init for a brms model
}
