% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_density_distribution_plots.R
\name{prior_densities}
\alias{prior_densities}
\title{Create a plot of the prior density distributions of modeled parameters
from brmsfit with sample_prior = "only".}
\usage{
prior_densities(model, title_label = "Prior Density Plots")
}
\arguments{
\item{model}{brmsfit of sample_prior = "only".}

\item{title_label}{string. Plot title name. (Default = "Prior Density Plot")}
}
\value{
ggplot object.
}
\description{
Create a plot of the prior density distributions of modeled parameters
from brmsfit with sample_prior = "only".
}
