% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_density_distribution_plots.R
\name{prior_posterior_densities}
\alias{prior_posterior_densities}
\title{Create a plot of the prior & posterior density distributions of modeled
parameters from brmsfit}
\usage{
prior_posterior_densities(model, title_label = "Prior Posterior Density Plots")
}
\arguments{
\item{model}{brmsfit.}

\item{title_label}{string. Plot title name. (Default = "Prior Posterior Density Plot")}
}
\value{
ggplot object.
}
\description{
Create a plot of the prior & posterior density distributions of modeled
parameters from brmsfit
}
