% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{rstan_default_init}
\alias{rstan_default_init}
\title{Helper Function to Create the Default \code{rstan} scalar init}
\usage{
rstan_default_init(lb = NULL, ub = NULL, dim = 0)
}
\arguments{
\item{lb}{\code{numeric} lower bound for parameter}

\item{ub}{\code{numeric} upper bound for parameter}

\item{dim}{\code{numeric} dimension of parameter.}
}
\value{
function that return the default brms initial value for a parameter.
If \code{dim=0}, then it will be a numeric scalar, if \code{dim=1} or greater, than
return an array with the given dimension.
}
\description{
By default, \pkg{rstan} will initialize parameters uniformly at
random in the range (-2, 2), on the unconstrained scale. Description of how
stan transforms parameters to satisfy constraints is described in the stan
documentation
\link{https://mc-stan.org/docs/reference-manual/variable-transforms.html}

This helper is especially useful for running models using the \pkg{cmdstanr}
backend, which requires all parameters (including distributional) parameters
to be initialized.
}
\examples{
\dontrun{

# Explicitly set the default initialization for the distributional parameter
# 'sigma' when family=gaussian().
init <- BayesPharma::sigmoid_antagonist_init(
  sigma = BayesPharma::rstan_default_scalar_init(lb = 0))

}

}
\seealso{
\link[rstan:stan]{rstan::stan}
}
