% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-function.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(ac50, hill, top, bottom, log_dose)
}
\arguments{
\item{ac50}{\code{numeric}, the \code{log_dose} of half maximal activity}

\item{hill}{\code{numeric}, slope of curve at the \code{ac50} units of
response/log_dose}

\item{top}{\code{numeric}, value of maximal response for positive slope,
this occurs when \code{log_dose = Inf}, with negative slope when
\code{log_dose = -Inf}}

\item{bottom}{\code{numeric}, value of minimal response for positive
slope, this occurs when log_dose = -Inf, with negative slope
when log_dose = Inf}

\item{log_dose}{\code{numeric}, where to evaluate the the response where
the log is base 10.}
}
\value{
\code{numeric}, response given the \code{log_dose} and parameters
}
\description{
Functional form for the sigmoid model.
}
\examples{
\dontrun{
 # Generate an agonist curve with an ec50 of 1 μM with the response
 # normalized to [0, 1] over the range 100 nM to 10 μM with half-log steps
 data <- data.frame(
   log_dose = seq(-7, -5, length.out = 5)) |>
     dplyr::mutate(
       response = stats::rnorm(
         n = length(log_dose),
         mean = BayesPharma::sigmoid(
           ac50 = -6,
           hill = 1,
           top = 1,
           bottom = 0,
           log_dose = log_dose),
       sd = .2))
}

}
\seealso{
\link{sigmoid_model}
}
