% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-init.R
\name{sigmoid_agonist_init}
\alias{sigmoid_agonist_init}
\title{Create initialization for the agonist sigmoid model}
\usage{
sigmoid_agonist_init(ec50 = -6, hill = 1, top = 1, bottom = 0)
}
\arguments{
\item{ec50}{\code{numeric} or \code{numeric} returning function units
log_dose (Default: -6, corresponding to 1 μM)}

\item{hill}{numeric or numeric returning function with units
response/log_dose (default = 1)}

\item{top}{numeric units of top (default = 1).}

\item{bottom}{numeric units of bottom (default = 0).}
}
\value{
input for \code{\link[brms]{brm}(init = ...)}
}
\description{
Creating initial values for an agonist sigmoid model
 parameters that can be passed to the \code{\link{sigmoid_agoninst_model}}.
}
\examples{
\dontrun{
#Consider an activator that has a max response around 50\%, EC50 is estimated
#to be around 1 nM, minimum response is known to be 0.
init <- BayesPharma::sigmoid_agonist_init(
  ec50 = -9,
  top = 0.5)
}
}
