% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-init.R
\name{sigmoid_antagoinst_init}
\alias{sigmoid_antagoinst_init}
\title{Create initialization for the antagonist sigmoid model}
\usage{
sigmoid_antagoinst_init(ic50 = -9, hill = -1, top = 100, bottom = 0)
}
\arguments{
\item{ic50}{numeric or numeric returning function units log_dose
(default = -9)}

\item{hill}{numeric or numeric returning function with units
response/log_dose (default = -1)}

\item{top}{numeric units of top (default = 100).}

\item{bottom}{numeric units of bottom (default = 0).}
}
\value{
input for brms::brm(init = ...)
}
\description{
Creating initial values for the sigmoid model parameters that
  can be passed to the `sigmoid_antagonist_model`
}
\examples{
\dontrun{
#Consider an inhibitor that has a min response around 50\%, IC50 is estimated
#to be around 1 uM, maximum response is known to be 1000,
init <- BayesPharma::sigmoid_antagonist_init(
  ec50 = -6,
  bottom = 50)
}
}
