% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-run.R
\name{sigmoid_antagonist_model}
\alias{sigmoid_antagonist_model}
\title{Run Bayesian Sigmoid Antagonist Model}
\usage{
sigmoid_antagonist_model(
  data,
  formula = sigmoid_antagonist_formula(),
  prior = sigmoid_antagonist_prior(),
  init = sigmoid_antagonist_init(),
  iter = 8000,
  control = list(adapt_delta = 0.99),
  stanvar_function = sigmoid_stanvar,
  expose_functions = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame of experimental data.  must contain columns
\code{sponse} and any predictors specified in the formula.}

\item{formula}{brmsformula object. To create a dose-response
brmsformula, (default: \code{BayesPharma::sigmoid__formula()}.}

\item{prior}{brmspriors data.frame for \code{ec50}, \code{hill},
\code{top}, and \code{bottom}.  Use one of the priors functions
provided to create priors to use here.  (default:
\code{BayesPharma::sigmoid_antagonist_prior()}}

\item{init}{initial values of the parameters being modeled (default
= \code{BayesPharma::sigmoid_antagonist_init()}}

\item{iter}{number of iterations the model runs. Increasing iter
can help with model convergence (default: 8000).}

\item{control}{a named list of parameters to control the sampler's
behavior.  Adding \code{max_treedepth} and giving a greater
value than 10 can improve model convergence (default:
\code{list(adapt_delta = 0.99)}).}

\item{stanvar_function}{stan code for the model (default:
\code{BayesPharma::sigmoid_stanvar})}

\item{expose_functions}{boolean. Expose the BayesPharma functions for the
model [default: TRUE].}

\item{...}{additional arguments passed to \code{brms::brm}}
}
\value{
\code{brmsfit} object
}
\description{
For additional information on additional function arguments, reference:
  <https://paul-buerkner.github.io/brms/reference/brm.html>
  or
  <https://rdrr.io/cran/rstan/man/stan.html>
}
\examples{
\dontrun{
  sigmoid_antagonist_model(data,
   formula = sigmoid_antagonist_formula(predictors = 0 + drug))
}
}
