% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-prior.R
\name{sigmoid_antagonist_prior}
\alias{sigmoid_antagonist_prior}
\title{Default priors for the sigmoid antagonist model}
\usage{
sigmoid_antagonist_prior(
  ic50 = brms::prior(prior = normal(-6, 2.5), nlpar = "ic50"),
  hill = brms::prior(prior = normal(-1, 1), nlpar = "hill", ub = 0.01),
  top = brms::prior(prior = normal(1, 0.5), nlpar = "top"),
  bottom = brms::prior(prior = normal(0, 0.5), nlpar = "bottom"),
  ...
)
}
\arguments{
\item{ic50}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
\code{ic50} parameter. (Default: \code{normal(-6, 2.5)} where the mean
\code{-6} corresponds to a concentration of 1e-6 or 1 μM. Setting
\code{ic50} to a numeric value constrains it to a constant value.}

\item{hill}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
\code{hill} parameter. (Default: \code{normal(-1, 1)} upper bounded by
\code{0.1}).}

\item{top}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
\code{top} parameter. (Default: \code{normal(1, 0.5)}. Setting \code{top}
to a numeric value constrains it to a constant value.}

\item{bottom}{\code{\link[brms:set_prior]{brms::brmsprior()}} or \code{numeric}. Prior for the
bottom parameter. (Default: \code{normal(0, 0.5)}. Setting \code{bottom} to
a numeric value constrains it to a constant value.}

\item{...}{additional \code{\link[brms:set_prior]{brms::brmsprior()}} objects.}
}
\value{
\code{\link[brms:set_prior]{brms::brmsprior()}} \code{data.frame}
}
\description{
Creates a \code{data.frame} of priors for the sigmoid antagonist
model

If the function arguments ic50, hill, top, bottom = NULL, default
normal distribution priors will be set.

If you would like to set a parameter as a constant, enter a numeric value for
the function argument.

For other distribution options, reference
\url{http://mc-stan.org/rstanarm/reference/priors.html#arguments}
}
\examples{
\dontrun{
# Consider an activator that has a max response around 50\%, EC50 is estimated
# to be around 1 μM, and minimum response is known to be 0.
priors <- sigmoid_antagonist_prior(
  ic50 = brms::prior(prior = normal(-6, 0.5), nlpar = "ic50"),
  top = brms::prior(prior = normal(0.5, 0.5), nlpar = "top"),
  bottom = 0)
}
}
