% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-formula.R
\name{sigmoid_formula}
\alias{sigmoid_formula}
\title{Create a sigmoid formula for the brms model}
\usage{
sigmoid_formula(predictors = 1, ...)
}
\arguments{
\item{predictors}{Additional formula objects to specify predictors of
non-linear parameters. i.e. what perturbations/experimental differences
should be modeled separately? (Default: 1) should a random effect be taken
into consideration? i.e. cell number, plate number, etc.}
}
\value{
brmsformula
}
\description{
set-up a sigmoid dose response model formula to define a
  non-linear model or multilevel non-linear model for `ec50`, `hill`, `top`,
  and, `bottom` for use in Bayesian_model and in the BRMS package.
}
\examples{
\dontrun{
  Consider observations made using 4 different drugs and the column header
  containing the labels for the 4 different drugs is `predictors`.
  
    `BayesPharma::sigmoid_formula(predictors = 0 + predictors)`

  Consider that the cell_ID was recorded and the noise from the cell_ID is to
  be accounted for. 
  
    `BayesPharma::sigmoid_formula(predictors = 0 + predictors + (1|cell_ID))`
}

}
