% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-function.R
\name{sigmoid_hill_to_slope}
\alias{sigmoid_hill_to_slope}
\title{Convert Hill to Slope Parameters a the Sigmoid Functional Form}
\usage{
sigmoid_hill_to_slope(hill, top, bottom)
}
\arguments{
\item{hill}{\code{numeric} the hill coefficient for the \code{\link[=sigmoid]{sigmoid()}}
function.}

\item{top}{\code{numeric} the top parameter for the \code{\link[=sigmoid]{sigmoid()}}
function.}

\item{bottom}{\code{numeric} the bottom parameter for the
\code{\link[=sigmoid]{sigmoid()}} function.}
}
\value{
\code{numeric} the slope of the \code{\link[=sigmoid]{sigmoid()}} function at
the ac50.
}
\description{
The \code{hill} parameter controls the slope at the ac50 (\code{slope}) but
the \code{slope} also depends on the \code{top} and \code{bottom} parameters.
This helper function facilitates computing the \code{slope} from the
parameters.
}
\details{
To prove the this function is correct, consider the derivative of the
response of the \code{\link[=sigmoid]{sigmoid()}} function with respect to the
\code{log_dose}:
\preformatted{
  d(response)/d(log_dose) =
    hill * log(10) * (top - bottom) * 10 ^ ((ac50 - log_dose) * hill) /
    ((10 ^ ((ac50 - log_dose) * hill) + 1) ^ 2)}
Evaluating this derivative at log_dose == ac50 simplifies the exponents:
\preformatted{
  slope = hill * log(10) * (top - bottom) / (1 + 1) ^ 2)
        = hill * log(10) * (top - bottom) / 4}
}
\seealso{
\code{\link[=sigmoid]{sigmoid()}} \code{\link[=sigmoid_slope_to_hill]{sigmoid_slope_to_hill()}}
}
