% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-function.R
\name{sigmoid_slope_to_hill}
\alias{sigmoid_slope_to_hill}
\title{For the sigmoid functional form, convert slope to the hill parameter}
\usage{
sigmoid_slope_to_hill(slope, top, bottom)
}
\arguments{
\item{slope}{\code{numeric} the slope of the \code{\link[=sigmoid]{sigmoid()}} at the
ac50}

\item{top}{\code{numeric} the top parameter for the \code{\link[=sigmoid]{sigmoid()}}
function.}

\item{bottom}{\code{numeric} the bottom parameter for the
\code{\link[=sigmoid]{sigmoid()}} function.}
}
\value{
\code{numeric} the hill coefficient of the \code{\link[=sigmoid]{sigmoid()}}
function at.
}
\description{
The \code{hill} parameter controls the slope at the ac50 (\code{slope}) but
the \code{slope} also depends on the \code{top} and \code{bottom} parameters.
This helper function facilitates computing the \code{hill} from the
parameters.
}
\details{
To prove the this function is correct, we will re-arrange the equation
relating the \code{slope} to the \code{hill} parameters
\preformatted{
  hill * log(10) * (top - bottom) / 4 = slope
  hill = slope * 4 / (log10 * (top - bottom))}
}
\seealso{
\code{\link[=sigmoid]{sigmoid()}} \code{\link[=sigmoid_hill_to_slope]{sigmoid_hill_to_slope()}}
}
