% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-stanvars.R
\name{sigmoid_stanvar}
\alias{sigmoid_stanvar}
\title{Stan Code for the Sigmoid Model.}
\usage{
sigmoid_stanvar()
}
\description{
Stan Code for the Sigmoid Model.
}
\note{
stanvar script of a sigmoid dose response equation.  For log dose
greater than negative infinity (dose greater than 0), the sigmoid dose
response equation is used. \code{ac50} is the half maximal response, hill is the
slope, top is the maximum response, and bottom is the minimum response. For
log dose equal to negative infinity (dose equal to 0), if hill is positive,
then the bottom value is returned. If hill is negative, then the top value
is returned.
}
