% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-run.R
\name{tQ_formula}
\alias{tQ_formula}
\title{Define formula for the tQ enzyme kinetics model}
\usage{
tQ_formula(predictors = 1, ...)
}
\arguments{
\item{predictors}{predictors to use for kcat and kM}

\item{multiple_perturbations}{(default FALSE)}
}
\description{
Define formula for the tQ enzyme kinetics model
}
