% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-stanvars.R
\name{tQ_genquant_stanvar}
\alias{tQ_genquant_stanvar}
\title{Stan Code for the tQ Model Generated Quantities}
\usage{
tQ_genquant_stanvar()
}
\description{
If only the substrate concentration is varied, it is not generally possible
to fit both \code{kcat} and \code{kM}. However, it is possible to fit the ratio
\code{kcat/kM}. Including this \link[rstan:stan]{rstan::stan} code will generate the samples for
\code{kcat/kM}.
}
\references{
Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation:
Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7,
17018 (2017). https://doi.org/10.1038/s41598-017-17072-z
}
\seealso{
\link{tQ_model}
}
