% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-init.R
\name{tQ_init}
\alias{tQ_init}
\title{Initialize parameter values for the tQ enzyme kinetic model}
\usage{
tQ_init(kcat = 4, kM = 4)
}
\arguments{
\item{kcat}{\code{numeric} initial value for \code{kcat} parameter in the tQ model}

\item{kM}{\code{numeric} initial value for the \code{kM} parameter in the tQ model}
}
\value{
input for \verb{[model_tQ](init = ...)} parameter.
}
\description{
Initialize parameter values for the tQ enzyme kinetic model
}
\seealso{
\link{tQ_model}, \link{tQ_formula}, and \link{tQ_prior}
}
