% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-run.R
\name{tQ_model}
\alias{tQ_model}
\title{BRMS model for the tQ enzyme kinetics model}
\usage{
tQ_model(
  data,
  formula = mm_formula(),
  prior = NULL,
  init = 0,
  iter = 8000,
  control = list(adapt_delta = 0.99),
  ...
)
}
\description{
BRMS model for the tQ enzyme kinetics model
}
