% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-prior.R
\name{tQ_prior}
\alias{tQ_prior}
\title{Define priors for the tQ enzyme kinetics model}
\usage{
tQ_prior(kcat = NULL, kM = NULL)
}
\arguments{
\item{kcat}{prior for `r kcat` parameter (Default: NULL). Given a numeric value,
it will be used as a constant and not estimated.}

\item{kM}{prior for `r kM` parameter (Default: NULL). Given a numeric value,
it will be used as a constant and not estimated.}
}
\description{
Default priors are gamma(4, 1) for both `r kcat` and `r kM`. We use the gamma
distribution because it is naturally lower bounded by 0. The first parameter
is the shape, `r alpha=4`, and the second is the rate, `r beta=1`. The mean
of gamma distributions is `r alpha/beta` and the variance is
`r alpha/beta^2`.
}
