% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-run.R
\docType{data}
\name{tQ_stanvars}
\alias{tQ_stanvars}
\title{stanvars for the tQ enzyme kinetics model}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
Pass to calls to `brms::brm`
}
\description{
stanvars for the tQ enzyme kinetics model
}
\keyword{datasets}
