% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{traceplot}
\alias{traceplot}
\title{Create a plot of MCMC chains and sample value at each MCMC iteration}
\usage{
traceplot(
  model,
  predictors_col_name = "predictors",
  half_max_label = NULL,
  title = NULL
)
}
\arguments{
\item{model}{brmsfit}

\item{predictors_col_name}{string. Name of the column with the perturbations
that each parameter is being estimated for.}

\item{half_max_label}{string. Label for the half maximal that fits the type of
experiment that was done (i.e. ec50, ic50, ed50, id50, ld50, etc.).}

\item{title}{string for the plot title (default = NULL)}
}
\value{
ggplot object
}
\description{
Create a plot of MCMC chains and sample value at each MCMC iteration
}
