% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{add_loo_criterion}
\alias{add_loo_criterion}
\title{Add model fit loo criterion to model objects}
\usage{
add_loo_criterion(model = <model>,
                    model_name = <model_name>)
}
\arguments{
\item{model}{brmsfit model.}

\item{model_name}{variable name given to the model.}

\item{...}{extra arguments passed on to `brms::add_criterion`}
}
\value{
brmsfit model with loo criterion.
}
\description{
This is required to use loo (leave one out) compare to compare
  multiple model fits against each other.
}
\examples{
  add_loo_criterion(model = my_dr_model,
                    model = my_dr_model_1)

}
