% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{add_loo_criterion}
\alias{add_loo_criterion}
\title{Add model fit loo criterion to model objects}
\usage{
add_loo_criterion(model, model_name, ...)
}
\arguments{
\item{model}{brmsfit model.}

\item{model_name}{variable name given to the model.}

\item{...}{extra arguments passed on to `brms::add_criterion`}

\item{n_params}{numeric value. The number of params being estimated by the
model. i.e. how many parameters have non-constant priors? in the case of dose
response cureves, are all 4 parameters being predicted by the model?}
}
\value{
brmsfit object with loo criterion.

brmsfit model with loo criterion.
}
\description{
This is required to use loo (leave one out) compare to compare
  multiple model fits against each other.
}
\examples{
  add_loo_criterion(model = my_dr_model,
                    model = my_dr_model_1)

}
