% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_density_distribution_plots.R
\name{basic_stats}
\alias{basic_stats}
\title{Displays a tibble of mean, median, standard deviation,
and confidence intervals}
\usage{
basic_stats(
  model,
  predictors_col_name = "predictors",
  half_max_label = NULL,
  l_ci = 0.025,
  u_ci = 0.975
)
}
\arguments{
\item{model}{brmsfit
#'@param predictors_col_name string. Name of the column with the perturbations
that each parameter is being estimated for.}

\item{half_max_label}{string. Label for the half maximal that fits the type of
experiment that was done (i.e. ec50, ic50, ed50, id50, ld50, etc.).}

\item{l_ci}{decimal of the lower confidence interval (default = 0.025)}

\item{u_ci}{decimal of the upper confidence interval (default = 0.975)}
}
\value{
tibble::tibble that is required for the 'posterior_densities'
function
}
\description{
Displays a tibble of mean, median, standard deviation,
and confidence intervals
}
