% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-basic-stats.R
\name{basic_stats}
\alias{basic_stats}
\title{Give Basic statistical information about the model results}
\usage{
basic_stats(
  model,
  predictors_col_name = "_Intercept",
  half_max_label = "ec50",
  l_ci = 0.025,
  u_ci = 0.975
)
}
\arguments{
\item{model}{\code{brms::brmsfit} model}

\item{predictors_col_name}{\code{string} expression for predictors column
in the input \code{data.frame}. Predictors are the perturbations tested
during the experiment (i.e. Drug, Temperature, etc.). [default =
"_Intercept"]}

\item{half_max_label}{\code{string} of the label for the half maximal that
fits the type of experiment that was done (i.e. ec50, ic50,
ed50, id50, ld50, etc.).}

\item{l_ci}{\code{numeric} unit of the lower confidence interval (default
= 0.025)}

\item{u_ci}{\code{numeric} unit of the upper confidence interval (default
= 0.975)}
}
\value{
\code{tibble::tibble} object.
}
\description{
Return a \code{data.frame} containing summary statistics of
    \code{brms::brmsfit} model.  The summary statistics included are
    mean, median, standard deviation, lower confidence interval, and upper
    confidence interval.
}
\examples{
\dontrun{
  BayesPharma::basic_stats(
    model = my_sigmoid_model,
    predictors_col_name = "predictors",
    half_max_label = "ic50",
    l_ci = 0.025,
    u_ci = 0.975)
}
}
