% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data_column_names.R
\name{change_col_names}
\alias{change_col_names}
\title{Change column names to work with BayesPharma functions}
\usage{
change_col_names(
  data,
  response_col_name,
  log_dose_col_name,
  predictors_col_name
)
}
\arguments{
\item{response_col_name}{string. Name of current column name for response.
if column name is 'response' then use NULL.}

\item{log_dose_col_name}{string. Name of current column name for log_dose.
if column name is 'log_dose' or log_dose hasn't been calculated yet, then use NULL.
to calculate log_dose, use the calc log_dose function.}

\item{predictors_col_name}{string. Name of current column name for perturbation
name (i.e. drug, temperature, etc.). if column name is 'predictors' or there
is only one perturbation, then use NULL.}
}
\value{
tibble or data.frame.
}
\description{
Change column names to work with BayesPharma functions
}
