% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data_column_names.R
\name{change_col_names}
\alias{change_col_names}
\title{Change column names to be compatible with BayesPharma functions}
\usage{
data %>% change_col_names(
        response_col_name = <response_col_name>,
        log_dose_col_name = <log_dose_col_name>,
        predictors_col_name = <predictors_col_name>)
}
\arguments{
\item{response_col_name}{expression for response column in the input data.frame (default = NULL)}

\item{log_dose_col_name}{expression for log_dose column in the input data.frame (default = NULL)
if log_dose hasn't been calculated yet, calculate log_dose, use the calculate_log_dose function.}

\item{predictors_col_name}{expression for predictors column in the input data.frame (default = NULL)}
}
\value{
input data.frame with new column header names for the response column (`response`),
  log_dose column (`log_dose`), and predictors column (`predictors`).
}
\description{
Given the response, log_dose, and predictors column of a data.frame,
  change the column names to `response`, `log_dose`, and `predictors` in the data.frame and return it.
}
\examples{
# consider the response column name is `normalized_measurement`, log dose column name is `log_10_dose`,
# and predictors column name is `Drug`.
data \%>\% change_col_names(response_col_name = normalized_measurement,
                          log_dose_col_name = log_10_dose,
                          predictors_col_name = Drug)

}
