% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data_column_names.R
\name{change_col_names}
\alias{change_col_names}
\title{Change column names to be compatible with BayesPharma functions}
\usage{
change_col_names(data, response_col_name = NULL, log_dose_col_name = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{response_col_name}{expression for response column in the input
data.frame (default = NULL).}

\item{log_dose_col_name}{expression for log_dose column in the input
data.frame (default = NULL). if log_dose hasn't been calculated yet,
calculate log_dose, use the \code{calculate_log_dose} function.}
}
\value{
input data.frame with new column header names for the response
   column (`response`) and log_dose column (`log_dose`).
}
\description{
Given the response, log_dose, and predictors column of a
  data.frame, change the column names to `response` and `log_dose` in the
  data.frame and return it.
}
\examples{
\dontrun{
# consider the response column name is `normalized_measurement` and log dose
# column name is `log_10_dose`.
data \%>\% BayesPharma::change_col_names(
  response_col_name = normalized_measurement,
  log_dose_col_name = log_10_dose)
}
}
