% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dr_brm_model.R
\name{dr_formula}
\alias{dr_formula}
\title{Create a formula for a multiple different perturbations/experimental variables
for the brms model}
\usage{
dr_formula(multiple_perturbations = FALSE, predictors = 0 + predictors, ...)
}
\arguments{
\item{multiple_perturbations}{TRUE/FALSE. If FALSE, the model will produce a
singular estimate for each parameter. If TRUE, the parameter will produce
parameter estimates for each perturbation.}

\item{predictors}{Additional formula objects to specify predictors of non-linear
parameters. i.e. what perturbations/experimental differences should be modeled
separately? (Default: 0 + predictors) should a random effect be taken into consideration?
i.e. cell number, plate number, etc.}
}
\value{
brmsformula
}
\description{
Create a formula for a multiple different perturbations/experimental variables
for the brms model
}
