% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-init.R
\name{dr_inits}
\alias{dr_inits}
\title{Create initialization for the sigmoid model}
\usage{
dr_inits(
  ec50 = -9,
  hill = TRUE,
  inhibitor = TRUE,
  top = 100,
  bottom = 0,
  chains = 4
)
}
\arguments{
\item{ec50}{numeric or numeric returning function units log_dose
(default = -9)}

\item{hill}{TRUE, FALSE, or numeric units, TRUE/FALSE sets hill to an
initial value of 1/-1 or numeric units can be assigned to hill
(default = TRUE).}

\item{inhibitor}{TRUE/FALSE value that determines if hill is a positive or
negative slope. If TRUE, the initial condition will be hill = -1. If FALSE,
hill = 1 (default = TRUE).}

\item{top}{numeric units of top (default = 100).}

\item{bottom}{numeric units of bottom (default = 0).}

\item{chains}{numeric units of Markov chains that will be used in brmsfit
(default = 4).}
}
\value{
list of lists.
}
\description{
Creating initial values for the sigmoid model parameters that
  can be passed to the `dr_model`
}
\examples{
\dontrun{
Consider an activator that has a max response around 50\%, EC50 is estimated
to be around 1 uM, minimum response is known to be 0, and 4 MCMC chains will
be used by the model.

inits <- BayesPharma::dr_inits(
  ec50 = -6,
  hill = TRUE,
  inhibitor = FALSE,
  top = 50,
  bottom = 0,
  chains = 4)
}
}
