% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dr_brm_model.R
\name{dr_model}
\alias{dr_model}
\title{Run Bayesian Regression Model using Stan}
\usage{
dr_model(
  data,
  formula,
  priors = NULL,
  init = 0,
  iter = 8000,
  control = list(adapt_delta = 0.99),
  stanvar_function = dr_stanvar,
  ...
)
}
\arguments{
\item{data}{data.frame of experimental data.
must contain columns \code{response} and any predictors specified in
the formula.}

\item{formula}{brmsformula object. To create a dose-response brmsformula,
use the \code{dr_formula} function.}

\item{priors}{brmspriors data.frame for ec50, hill, top, and bottom.
Use one of the priors functions provided to create priors to use here.}

\item{init}{list of lists, numeric value, or "random" for the initial values
of the parameters being modeled (default = 0).}

\item{iter}{number of iterations the model runs. Increasing iter can help
with model convergence (default = 8000).}

\item{control}{a named list of parameters to control the sampler's behavior.
Adding max_treedepth and giving a greater value than 10 can improve model
convergence (default = list(adapt_delta = 0.99)).}

\item{stanvar_function}{stan code for the model.}
}
\value{
brmsfit model
}
\description{
For additional information on additional function arguments, reference:
  https://paul-buerkner.github.io/brms/reference/brm.html
  or
  https://rdrr.io/cran/rstan/man/stan.html
}
\examples{
\dontrun{
  dr_model(data,
   formula = dr_formula(predictors = 0 + predictors),
   priors = dr_priors(),
   inits = dr_inits(),
   iter = 8000,
   control = list(adapt_delta = 0.99),
   stanvar_function = dr_stanvar)
}
}
