% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dr_brm_model.R
\docType{data}
\name{dr_stanvar}
\alias{dr_stanvar}
\title{stanvar function}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
dr_stanvar
}
\description{
stanvar function
}
\keyword{datasets}
