% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-drc.R
\name{drc_model}
\alias{drc_model}
\title{Wrapper for DRC dose response models}
\usage{
drc_model(
  data,
  formula = response ~ log_dose,
  fct = drc::L.4(fixed = c(NA, NA, NA, NA), names = c("hill", "bottom", "top", "ic50")),
  ...
)
}
\arguments{
\item{data}{\code{data.frame}}

\item{formula}{\code{formula}. Default: \code{response ~ log_dose}}

\item{fct}{Non-linear link function. Default is the 4 parameter
log-linear sigmoid equation, where all parameters are to be
estimated. To fix a subset them, set the parameters to constant
values.}

\item{...}{additional arguments passed to \code{drc::drm}}
}
\value{
An object of (S3) class 'drc'.
}
\description{
Wrapper for DRC dose response models
}
