% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-testthat.R
\name{expect_gg}
\alias{expect_gg}
\title{Check if an Object is a \link{ggplot2} Plot}
\usage{
expect_gg(x)
}
\arguments{
\item{x}{\link[ggplot2:ggplot]{ggplot2::ggplot} object}
}
\description{
Use in tests to check if an object is a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\note{
borrowed from \code{bayesplot:::expect_gg}
}
