% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-function.R
\name{growth_richards}
\alias{growth_richards}
\title{Richards growth function}
\usage{
growth_richards(K, K0, rate, lambda, nu, time)
}
\arguments{
\item{K}{\code{numeric}, the carrying capacity}

\item{K0}{\code{numeric}, the baseline response}

\item{rate}{\code{numeric}, maximum growth rate}

\item{lambda}{\code{numeric}, time line through the maximum growth rate
crosses zero}

\item{nu}{\code{numeric}, asymmetry in growth before and after the inflection
point}

\item{time}{\code{numeric}, time point at which to evaluate the response.}
}
\value{
numeric, response given the time and parameters

  response(time) =
    K0 + (K - K0)/(
      1 + nu * exp(
        1 + nu + rate/(K - K0) * (1 + nu)^(1 + 1/nu) *
        (lambda - time))) ^ (1/nu)
}
\description{
Functional form for the Richards growth model.
}
\examples{
\dontrun{

 # Generate Richards growth curve
 data <- data.frame(
   time = seq(0, 2, length.out = 101)) |>
     dplyr::mutate(
       response = stats::rnorm(
         n = length(time),
         mean = BayesPharma::growth_richards(
           K = 1,
           K0 = 0,
           rate = 2,
           lambda = 0.5,
           nu = 2,
           time = time),
       sd = .2))
}

}
