% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-formula.R
\name{growth_richards_formula}
\alias{growth_richards_formula}
\title{Create a formula for the Richards growth model}
\usage{
growth_richards_formula(
  treatment_variable = "time",
  treatment_units = "hours",
  response_variable = "response",
  response_units = NULL,
  predictors = 1,
  ...
)
}
\arguments{
\item{treatment_variable}{character variable representing time as a treatment
(Default: 'time')}

\item{treatment_units}{character the units of the time variable. (Default:
"seconds")}

\item{response_variable}{character variable representing the response to
treatment. (Default: 'response')}

\item{response_units}{character the units of the response. (Default: NULL)}

\item{predictors}{Additional formula objects to specify predictors
of non-linear parameters. i.e. what perturbations/experimental differences
should be modeled separately? (Default: 1) should a random effect be taken
into consideration? i.e. cell number, plate number, etc.}

\item{...}{additional arguments to \code{\link[brms]{brmsformula}}}
}
\value{
a \code{bpformula}, which is a subclass of
  \code{\link[brms]{brmsformula}} and can be passed to
  \code{\link{growth_richards_model}}.
}
\description{
set-up a Richards growth model formula to for use in
  \code{growth_richards_model} and in the BayesPharma package. The functional
  form is
  
    response ~ richards_growth(K, K0, rate, lambda, nu, time)
  
  The parameterization follows (Zwietering, 1990) and grofit:
  
    K      = carrying capacity, K = response(time = Inf). In the grofit
             package this is parameter is called "A", having the same units
             as the response
    K0     = initial population size K0 = response(time = 0). In the grofit
             package K0 is assumed to be zero, having the same units as the
             response.
    rate   = the maximum growth rate The grofit package calls this μ, having
             with of response/time
    nu     = how asymmetric the growth is before and after the inflection
             point
    lambda = the length of the lag-phase, defined by to be the time point at
             which the tangent through the growth curve crosses response =
             K0.
    See the vignettes(topic = "derive_growth_model", package = "BayesPharma")
}
\examples{
\dontrun{
  # Data has a string column drug_id with drug identifiers
  # Fit a separate model for each drug
  BayesPharma::growth_richards_formula(predictors = 0 + drug_id)

  # Data has a string column plate_id with plate identifiers
  # Estimate the change in response for each plate relative to a global
  # baseline.
  BayesPharma::growth_richards_formula(predictors = plate_id)

  # data has columns drug_id and plate_id
  # fit a multilevel model where the drug effect depends on the plate
  BayesPharma::growth_richards_formula(predictors = 0 + (drug_id|plate_id))
}

}
\seealso{
\code{\link[brms]{brmsformula}}, which this function wraps.
  \code{\link{growth_richards_model}} into which the result of this
  function can be passed.
}
