% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-init.R
\name{growth_richards_init}
\alias{growth_richards_init}
\title{Create Initialization for the Richards Growth Model}
\usage{
growth_richards_init(
  K = function() stats::runif(n = 1, min = 0.8, max = 1.2),
  K0 = function() stats::runif(n = 1, min = -0.2, max = 0.2),
  rate = function() stats::runif(n = 1, min = 0.8, max = 1.2),
  lambda = function() stats::runif(n = 1, min = 0.3, max = 0.7),
  nu = function() stats::runif(n = 1, min = 0.8, max = 1.2),
  ...
)
}
\arguments{
\item{K}{\code{numeric} or \code{numeric} returning \code{function}}

\item{K0}{\code{numeric} or \code{numeric} returning \code{function}}

\item{rate}{\code{numeric} or \code{numeric} returning \code{function}}

\item{lambda}{\code{numeric} or \code{numeric} returning \code{function}}

\item{nu}{\code{numeric} or \code{numeric} returning \code{function}
values range between [0, Inf], with 1 being standard sigmoid}

\item{...}{additional parameter initialization. Each named argument should be
a function that returns an \code{numeric} \code{array} of length \code{1}, see
\code{\link[=rstan_default_init]{rstan_default_init()}} to use the rstan default init.}
}
\value{
input for \verb{[brm][brms::brm](init = ...)}
}
\description{
Create initial values for Richards growth model parameters
that can be passed to the \link{growth_richards_model}.
}
\examples{
\dontrun{
init <- BayesPharma::growth_richards_init(
  A = 100,
  nu = 2)
}
}
\seealso{
\link{growth_richards_formula}, \link{growth_richards_prior}, and
\link{growth_richards_model}
}
