% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-stanvars.R
\name{growth_richards_stanvar}
\alias{growth_richards_stanvar}
\title{Stan Code for the Richards Growth Function.}
\usage{
growth_richards_stanvar()
}
\description{
Stan Code for the Richards Growth Function.
}
\note{
stanvar script of a Richard growth model response equation.
}
\examples{
\dontrun{
brms::brm(
  data = data.frame(
    response = ...,
    time = ...),
  formula = brms::brmsformula(
    response ~ growth_richards(K, K0, rate, lambda, nu, time),
    K + K0 + rate + lambda + nu ~ 1,
    nl = TRUE),
  prior = ...,
  init =  ...,
  stanvars = BayesPharma::growth_sigmoid_stanvar())
}

}
