% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-init.R
\name{growth_sigmoid_init}
\alias{growth_sigmoid_init}
\title{Create initialization for the sigmoid growth model}
\usage{
growth_sigmoid_init(K = 1, K0 = 0, rate = 1, lambda = 0.5)
}
\arguments{
\item{K}{\code{numeric} or \code{numeric} returning \code{function}}

\item{K0}{\code{numeric} or \code{numeric} returning \code{function}}

\item{rate}{\code{numeric} or \code{numeric} returning \code{function}}

\item{lambda}{\code{numeric} or \code{numeric} returning \code{function}
units of \code{time}}
}
\value{
input for \verb{[brm][brms::brm](init = ...)}
}
\description{
Creating initial values for sigmoid growth model
parameters that can be passed to the
.  \link{growth_sigmoid_model}.
}
\examples{
\dontrun{
init <- BayesPharma::growth_sigmoid_init(
  A = 100)
}
}
\seealso{
\link{growth_sigmoid_formula}, \link{growth_sigmoid_prior}, and
\link{growth_sigmoid_model}
}
