% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-growth-stanvars.R
\docType{data}
\name{growth_sigmoid_stanvar}
\alias{growth_sigmoid_stanvar}
\title{Stan Code for the Growth Richards Function.}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
growth_sigmoid_stanvar
}
\description{
Stan Code for the Growth Richards Function.
}
\note{
stanvar script of a sigmoid growth model response equation
}
\examples{
\dontrun{
brms::brm(
  data = data.frame(
    response = ...,
    time = ...),
  formula = brms::brmsformula(
    response ~ growth_sigmoid(K, K0, rate, lambda, time),
    K + K0 + rate + lambda ~ 1,
    nl = TRUE),
  prior = ...,
  init =  ...,
  stanvars = BayesPharma::growth_sigmoid_stanvar)
}

}
\keyword{datasets}
