% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-michaelis_menten-init.R
\name{michaelis_menten_init}
\alias{michaelis_menten_init}
\title{Initialize Parameter Values for the Michaelis Menten Enzyme Kinetic Model}
\usage{
michaelis_menten_init(
  kcat = function() runif(n = 1, min = 3, max = 5),
  kM = function() runif(n = 1, max = 3, min = 5),
  ...
)
}
\arguments{
\item{kcat}{\code{numeric} initial value for \code{kcat} parameter in the
Michaelis Menten model}

\item{kM}{\code{numeric} initial value for the \code{kM} parameter in the
Michaelis Menten model}

\item{...}{additional parameter initialization. Each named argument should be
a function that returns a \code{numeric} or \code{array} depending on the dimension
of the parameter, see \code{\link[=rstan_default_init]{rstan_default_init()}} to use the \pkg{rstan} default
init.}
}
\value{
input for \verb{[michaelis_menten_model](init = ...)} parameter.
}
\description{
Initialize Parameter Values for the Michaelis Menten Enzyme Kinetic Model
}
\seealso{
\link{michaelis_menten_model}, \link{michaelis_menten_formula}, and
\link{michaelis_menten_prior}
}
