% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-formula.R
\name{MuSyC_formula}
\alias{MuSyC_formula}
\title{Create a formula for the MuSyC synergy model}
\usage{
MuSyC_formula(predictors = 1, ...)
}
\arguments{
\item{predictors}{Additional formula objects to specify predictors of
non-linear parameters. i.e. what perturbations/experimental differences
should be modeled separately? (Default: 1) should a random effect be taken
into consideration? i.e. cell number, plate number, etc.}

\item{...}{additional arguments passed to `r brms::brmsformula`}
}
\value{
brmsformula
}
\description{
setup a defaul `r MuSyC` synergy model formula to predict
  the `r E0`, `r C1`, `r E1`, `r s1`, `r C2`, `r E2`, `r s2`, `r log10alpha`,
  and `r E3alpha` parameters.
}
\examples{
\dontrun{
  # Data has a string column `r drug_id` with drug identifiers
  # Fit a separate model for each drug
  BayesPharma::MuSyC_formula(predictors = 0 + drug_id)

  # Data has a string column `r plate_id` with plate identifiers
  # Estimate the change in response for each plate relative to a global
  # baseline.
  BayesPharma::MuSyC_formula(predictors = plate_id)

  # data has columns `r drug_id` and `r plate_id`
  # fit a multilevel model where the drug effect depends on the plate
  BayesPharma::MuSyC_formula(predictors = 0 + (drug_id|plate_id))
}

}
