% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-formula.R
\name{MuSyC_formula}
\alias{MuSyC_formula}
\title{Create a formula for the MuSyC synergy model}
\usage{
MuSyC_formula(
  treatment_1_variable = "logd1",
  treatment_1_units = "Log[Molar]",
  treatment_2_variable = "logd2",
  treatment_2_units = "Log[Molar]",
  response_variable = "response",
  response_units = NULL,
  predictors = 1,
  ...
)
}
\arguments{
\item{treatment_1_variable}{character variable representing the treatment.
(Default: 'logd1')}

\item{treatment_1_units}{character the units of the treatment. The default is
log base 10 of the molar concentration. (Default: "Log[Molar]")}

\item{treatment_2_variable}{character variable representing the treatment.
(Default: 'logd2')}

\item{response_variable}{character variable representing the response to
treatment. (Default: 'response')}

\item{response_units}{character the units of the response. (Default: NULL)}

\item{predictors}{Additional formula objects to specify predictors
of non-linear parameters. i.e. what perturbations/experimental
differences should be modeled separately? (Default: 1) should a
random effect be taken into consideration? i.e. cell number,
plate number, etc.}

\item{...}{additional arguments passed to \code{\link[brms]{brmsformula}}}
}
\value{
\code{\link[brms]{brmsformula}}
}
\description{
setup a default \code{MuSyC} synergy model formula to
    predict the \code{E0}, \code{C1}, \code{E1}, \code{s1},
    \code{C2}, \code{E2}, \code{s2}, \code{log10alpha}, and
    \code{E3alpha} parameters.
}
\examples{
\dontrun{
  # Data has a string column drug_id with drug identifiers
  # Fit a separate model for each drug
  BayesPharma::MuSyC_formula(predictors = 0 + drug_id)

  # Data has a string column plate_id with plate identifiers
  # Estimate the change in response for each plate relative to a global
  # baseline.
  BayesPharma::MuSyC_formula(predictors = plate_id)

  # data has columns drug_id and plate_id
  # fit a multilevel model where the drug effect depends on the plate
  BayesPharma::MuSyC_formula(predictors = 0 + (drug_id|plate_id))
}

}
