% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-run.R
\name{MuSyC_hi_to_si}
\alias{MuSyC_hi_to_si}
\title{Convert slope parametrization to exponent parametrization for drug i}
\usage{
MuSyC_hi_to_si(hi, Ci, E0, Ei)
}
\arguments{
\item{Ci}{the IC50 of drug i}

\item{E0}{the response with no treatments}

\item{Ei}{the response of infinite drug i and no other treatments}

\item{si}{slope of drug i at it's IC50 and doses of all other drugs are zero}
}
\value{
hi the exponent in the MuSyC equation for drug i
}
\description{
This can be used for setting priors and interpreting parameter estimates
see MuSyC_si_to_hi for details
}
