% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-init.R
\name{MuSyC_init}
\alias{MuSyC_init}
\title{Create initialization for the MuSyC model}
\usage{
MuSyC_init(
  logE0 = log(0.5),
  logC1 = 0,
  logE1 = log(0.5),
  h1 = 4,
  logC2 = 0,
  logE2 = log(0.5),
  h2 = 4,
  logE3 = log(0.5),
  logalpha = 0
)
}
\arguments{
\item{logE0}{`r numeric` or function returning `r array` of length `r 1`.
Initial value for the `r logE0` parameter. Default: `r log(0.5)`}

\item{logC1}{`r numeric` or function returning array of length 1. Initial
value for the `r logC1` parameter. Default: `r 0 = log(1)``}

\item{logE1}{`r numeric` or function returning array of length 1. Initial
value for the `r logE1` parameter. Default: log(0.5)}

\item{h1}{`r numeric` or function returning array of length 1. Initial value
for the `r h1` parameter. Default: `r 4 =
MuSyC_si_to_hi(si=1, Ci=1, E0=1, Ei=0.0)` the exponent for treatment `r 1`
corresponding to a slope of `r 1`}

\item{logC2}{`r numeric` or function returning array of length `r 1`. Initial
value for the `r logC2` parameter. Default: `r 0 = log(1)`}

\item{logE2}{`r numeric` or function returning array of length `r 1`. Initial
value for the `logE2` parameter. Default: `r log(0.5)`}

\item{h2}{`r numeric` or function returning array of length `r 1`. Initial
value for the `h2` parameter. Default: `r 4 =
MuSyC_si_to_hi(si=1, Ci=1, E0=1, Ei=0.0)` the exponent for treatment `r 1`
corresponding to a slope of `r 1`}

\item{logE3}{`r numeric` or function returning array of length 1. Initial
value for the `r logE3` parameter. Default: `r log(0.5)`}

\item{logalpha}{`r numeric` or function returning array of length 1. Initial
value for the `r logalpha` parameter. Default: `r 0 = log(1)`}
}
\value{
input for `r BayesPharma::model_MuSyC(init = ...)` parameter.
}
\description{
Initial values for the MuSyC model
}
\examples{
\dontrun{
  # default prior and init for MuSyC Model
  prior <- BayesPharma::MuSyC_prior()
  init <- BayesPharma::MuSyC_init()

  # Set prior and init for MuSyC model where both treatment 1 and treatment 2
  # are assumed to be full inhibitors
  BayesPharma::MuSyC_prior(logE0 = 0, E1 = -Inf, E2 = -Inf, E3 = -Inf)
  BayesPharma::MuSyC_init(E0 = 1, E1 = -Inf, E2 = -Inf, E3 = -Inf)
}
}
