% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-run.R
\name{MuSyC_model}
\alias{MuSyC_model}
\title{MuSyC Synergy Model}
\usage{
MuSyC_model(
  data,
  prior = MuSyC_prior(),
  init = MuSyC_init(),
  formula = MuSyC_formula(),
  control = list(adapt_delta = 0.99, max_treedepth = 12),
  stanvars = c(MuSyC_function_stanvar, MuSyC_genquant_stanvar),
  expose_functions = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame of experimental data.  must contain columns
\code{response} and any predictors specified in the formula.}

\item{prior}{\code{brmspriors} for \code{kcat} and \code{kM}
parameters. Use \code{BayesPharma::MuSyC_prior(...)} to create
the prior to use here.  Default:
\code{BayesPharma::MuSyC_prior()}.}

\item{init}{Initial values for the parameters, Use
\code{BayesPharma::MuSyC_init(...)} to create the init to use
here. Default \code{BayesPharma::MuSyC_init()}.}

\item{formula}{\code{brmsformula} object. To create a dose-response
\code{brmsformula}, use the \code{MuSyC_formula} function.}

\item{control}{a named list of parameters to control the sampler's
behavior.  Adding \code{max_treedepth} and giving a greater
value than 10 can improve model convergence (default:
\code{list(adapt_delta = 0.99)}).}

\item{stanvars}{\code{stanvars} code for the model. Default defines
the MuSyC function and genquant code.}

\item{expose_functions}{boolean. Expose the BayesPharma functions for the
model [default: TRUE].}

\item{...}{additional arguments passed to \code{brms::brm}}
}
\description{
Combined:
     sample1 =
}
\details{
nlp_logE0 = X_logE0 * b_logE0
                 _______   _______
                    = 1    = param

     nlp_logE1 = X_logE1 * b_logE0
                 _______   ______
                 =sample1
}
