% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-generate.R
\name{MuSyC_model_generate}
\alias{MuSyC_model_generate}
\title{Generate MuSyC Ed scores using a robust functional form
It should give the same results as the simple one, but
be more numerically stable}
\usage{
MuSyC_model_generate(
  logd1,
  logd2,
  logE0,
  logE1,
  logC1,
  h1,
  logE2,
  logC2,
  h2,
  logE3,
  logalpha
)
}
\arguments{
\item{logalpha}{}
}
\description{
Generate MuSyC Ed scores using a robust functional form
It should give the same results as the simple one, but
be more numerically stable
}
