% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-prior.R
\name{MuSyC_prior}
\alias{MuSyC_prior}
\title{Create a Prior for the MuSyC Synergy Model}
\usage{
MuSyC_prior(
  logE0 = brms::prior(prior = normal(-0.6931, 3), nlpar = "logE0", ub = 0),
  logE1 = brms::prior(prior = normal(-1.386, 3), nlpar = "logE1", ub = 0),
  logC1 = brms::prior(prior = normal(0, 3), nlpar = "logC1"),
  h1 = brms::prior(prior = normal(4, 3), nlpar = "h1", lb = 0.1),
  logE2 = brms::prior(prior = normal(-1.386, 3), nlpar = "logE2", ub = 0),
  logC2 = brms::prior(prior = normal(0, 3), nlpar = "logC2"),
  h2 = brms::prior(prior = normal(4, 3), nlpar = "h2", lb = 0.1),
  logE3 = brms::prior(prior = normal(-1.386, 3), nlpar = "logE3", ub = 0),
  logalpha = brms::prior(prior = normal(0, 0.5), nlpar = "logalpha"),
  ...
)
}
\arguments{
\item{logE0}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for the
log response with no treatment. (Default: \code{normal(log(0.5), 3)} lower
bounded by \code{0}).}

\item{logE1}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for the
full log response for treatment 1 and the treatment 2 is zero. (Default:
\code{normal(log(0.25), 3)} lower bounded by \code{0}).}

\item{logC1}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for the
log AC50 of treatment 1 and the dose of treatment 2 is zero. (Default:
\code{normal(log(1), 3)}).}

\item{h1}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for the
slope of treatment 1 at its AC50 and treatment 2 is zero in exponential
form. To converted the slope to exponential form, use
\code{\link[=MuSyC_si_to_hi]{MuSyC_si_to_hi()}}). In particular, let \code{mu = MuSyC_si_to_hi(si = 1, Ci = 1, E0 = 0, Ei = 0)}, then (Default:
\code{normal(mu, 3)} lower bounded by \code{.1}).}

\item{logE2}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for the
full log response for treatment 2 and the treatment 1 is zero. (Default:
\code{normal(log(0.25), 3)} lower bounded by \code{0}).}

\item{logC2}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for the
log AC50 of treatment 2. (Default: to \code{normal(log(1), 3)}).}

\item{h2}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior the slope of
treatment 2 at its AC50 and treatment 1 is zero in exponential form. To
converted the slope to exponential form, use \code{\link[=MuSyC_si_to_hi]{MuSyC_si_to_hi()}}).
In particular, let \code{mu = MuSyC_si_to_hi(si = 1, Ci = 1, E0 = 0, Ei = 0)}, then (Default: \code{normal(mu, 3)} lower bounded by \code{.1}).}

\item{logE3}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for the
combined full response for treatment 1 and treatment 2. (Default:
\code{normal(log(0.25), 3)} lower bounded by 0).}

\item{logalpha}{\code{numeric} or \code{\link[brms:set_prior]{brms::brmsprior()}}. Prior for
the log synergistic potency. Where values greater than \code{1} are
synergistic, values less than \code{1} are antagonistic. (Default:
\code{normal(log(0), 0.5)}.}

\item{...}{additional \code{\link[brms:set_prior]{brms::brmsprior()}} objects to add to the
prior.}
}
\value{
\code{\link[brms:set_prior]{brms::brmsprior()}} \code{data.frame}
}
\description{
Creating a \code{data.frame} of priors of the dose response
parameters required for the MuSyC model.

If you would like to set a parameter as a constant, enter a
numeric value for the function argument.  If you would like to set your own
distribution, for example, in the function arguments type
\verb{ec50 = brms::prior(prior = normal(<mu>, <sigma>), nlpar = 'ec50')}.
For other distribution options, reference \link[brms:set_prior]{brms::prior} and \href{https://github.com/stan-dev/stan/wiki/Prior-Choice-Recommendations}{Prior Choice Recommendations }
}
