% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-utils.R
\name{MuSyC_si_to_hi}
\alias{MuSyC_si_to_hi}
\title{Convert from slope to exponent parameterization for drug i}
\usage{
MuSyC_si_to_hi(si, Ci, E0, Ei)
}
\arguments{
\item{si}{\code{numeric} value of slope of drug i at its AC50 and doses of
all other drugs are zero}

\item{Ci}{\code{numeric} value of the AC50 of drug i}

\item{E0}{\code{numeric} value of the response with no treatments}

\item{Ei}{\code{numeric} value of the response of infinite drug i and no
other treatments}
}
\value{
hi \code{numeric} value of the exponent in the MuSyC equation for
drug i
}
\description{
This can be used for setting priors and interpreting parameter
estimates
}
\details{
Claim: When d1=0 and d2=C2 then the gradient of the response with
respect to d2 is the s2, symbolically d(Ed)/d(d2) = s2

\if{html}{\out{<div class="sourceCode">}}\preformatted{where s2 = h2 * (E0 + E2) / (4 * C2)
}\if{html}{\out{</div>}}

d(Ed)/d(d2)
= d/d(d2)
(C1^h1 * C2^h2 * E0 + C1^h1 * d2^h2 * E2) /
(C1^h1 * C2^h2      + C1^h1 * d2^h2)

Cancel the C1^h1 terms:
=  d/d(d2)
(C2^h2 * E0 + d2^h2 * E2) /
(C2^h2      + d2^h2)

Distribute the derivative across the terms in the numerator
=  E0 * C2^h2 * \link{d/d(d2) 1     / (C2^h2 + d2^h2)} +
E2         * \link{d/d(d2) d2^h2 / (C2^h2 + d2^h2)}

\if{html}{\out{<div class="sourceCode">}}\preformatted{=  E0 * C2^h2 * [h2 * d2^(h2-1) / (C2^h2 + d2^h2)^2] +
   E2 * [C2^h2 * h2 * d2^(h2-1) / (C2^h2 + d2^h2)^2]

=  (E0 + E2) * C2^h2 * h2 * d2^(h2-1)/(C2^h2 + d2^h2)^2
}\if{html}{\out{</div>}}

Evaluate at d2 = C2:
=  (E0 + E2) * h2 * C2^(2\emph{h2-1) / [4}C2^(2*h2))]
=  h2 * (E0 + E2) / (4 * C2)

[4\emph{C2^(2}h2))]: R:4\emph{C2\%5E(2}h2))
}
