% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MuSyC-stanvars.R
\docType{data}
\name{MuSyC_stanvar}
\alias{MuSyC_stanvar}
\title{Stan code for the MuSyC function.}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
MuSyC_stanvar
}
\description{
Stan code for the MuSyC function.
}
\note{
stanvar script of a MuSyC equation.
}
\keyword{datasets}
