% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_model_draws.R
\name{n_sample_draws}
\alias{n_sample_draws}
\title{Dataframe of posterior samples of modeled parameters from brmsfit}
\usage{
n_sample_draws(model, n = 100)
}
\arguments{
\item{model}{brmsfit.}

\item{n}{number of samples from the posterior distribution. (default n = 100)}
}
\value{
df that is required for the 'posterior_response_draws' function.
}
\description{
Dataframe of posterior samples of modeled parameters from brmsfit
}
