% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density-distirbution.R
\name{plot_density_distribution}
\alias{plot_density_distribution}
\title{Plot Density Distribution of Model Parameters}
\usage{
plot_density_distribution(
  model,
  pars = NULL,
  labeller = NULL,
  title_label = "Density Distributions"
)
}
\arguments{
\item{model}{`\link[brms:brmsfit-class]{brms::brmsfit} model.}

\item{pars}{parameters to choose (Default b_\if{html}{\out{<parameter>}}_Intercept')}

\item{labeller}{strip off the 'b_' and '_Intercept' from the parameter labels}

\item{title_label}{\code{character} of the plot title.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This function is useful to observe the distributions
of the priors set for the brmsfit model. By adding \code{sample_prior = "only"} to the sigmoid_model or a brmsfit model, the model samples only
from the prior distributions and can be observed using this plot function
to verify that the values intended to be covered are being included.
}
\examples{
\dontrun{
  plot_density_distribution(
    model = my_sigmoid_model)
}

}
