% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drc.R
\name{plot_drc}
\alias{plot_drc}
\title{Create plot of drc model}
\usage{
plot_drc(
  model,
  newdata = NULL,
  predict_args = list(interval = "prediction"),
  aes_mapping = ggplot2::aes(x = .data[["log_dose"]], y = .data[["response"]]),
  title = NULL
)
}
\arguments{
\item{model}{\code{drc::drc} model object generated with \code{\link[drc:drm]{drc::drm()}}}

\item{newdata}{\code{data.frame} where the predictions should be made.
Default: use the range of treatment values in the model object.}

\item{predict_args}{\code{list} of arguments passed to
\code{predict(model, ...)} including interval will generate ribbon of the
given type, see \code{\link[drc:predict.drc]{drc::predict.drc()}} for more details.}

\item{aes_mapping}{\verb{[aes][ggplot2::aes](...)} mapping columns of the
model data and new data to the plot. (Default:
\verb{ggplot2::[aes][ggplot2::aes](x = log_dose, y = response)})}

\item{title}{\code{character} giving the plot title}
}
\description{
Create plot of drc model
}
