% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-density.R
\name{plot_posterior_density}
\alias{plot_posterior_density}
\title{Plot Posterior Density Distributions of modeled parameters}
\usage{
plot_posterior_density(
  model,
  exclude_variables = c("__$", "lprior"),
  l_ci = 0.025,
  u_ci = 0.975,
  title_label = "Posterior Density Plots w/ Mean & 95\% CI"
)
}
\arguments{
\item{model}{\code{bpfit} model.}

\item{exclude_variables}{list of regular expressions to match what variables
to exclude.}

\item{l_ci}{\code{numeric} unit of the lower confidence interval}

\item{u_ci}{\code{numeric} unit of the upper confidence interval}

\item{title_label}{string of the plot title.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
plots the \code{\link[=bpfit]{bpfit()}} model density distribution of each
parameter and the mean, lower confidence interval, and upper confidence
interval.
}
\examples{
\dontrun{
  plot_posterior_density(
    model = my_sigmoid_model,
    l_ci = 0.025,
    u_ci = 0.975,
    title_label = "Posterior Density Plots with Mean and 95\% CI")
}
}
