% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prior_posterior_densities.R
\name{plot_prior_posterior_densities}
\alias{plot_prior_posterior_densities}
\title{Plot Prior and Posterior Marginal Distributions}
\usage{
plot_prior_posterior_densities(
  model,
  predictors_col_name = "_Intercept",
  half_max_label = "ec50",
  title_label = "Prior Posterior Density",
  ...
)
}
\arguments{
\item{model}{\code{bpfit} object resulting from fitting a BayesPharma model}

\item{predictors_col_name}{string expression for predictors column
in the input data.frame (default = "_Intercept"). Predictors
are the perturbations tested during the experiment (i.e. Drug,
Temperature, etc.).}

\item{half_max_label}{string of the label for the half maximal that
fits the type of experiment that was done (i.e. ec50, ic50,
ed50, id50, ld50, etc.).}

\item{title_label}{string of the plot title.}

\item{...}{further arguments passed to \code{\link[stats:update]{stats::update()}} to sample from the
prior}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generate a plot that shows for each model variable the
distribution of the marginal posterior uncertainty overlaid on the
distribution of the marginal prior uncertainty.
}
\examples{
\dontrun{
  plot_prior_posterior_densities(
    model = my_sigmoid_model,
    predictors_col_name = "predictors",
    half_max_response = "ic50",
    title_label = "Prior Posterior Density Plots")
}


}
