% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-synergy-checkerboard.R
\name{plot_synergy_checkerboard}
\alias{plot_synergy_checkerboard}
\title{Plot synergy checkerboard}
\usage{
plot_synergy_checkerboard(
  data,
  treatment_1_label = "Treatment 1",
  treatment_2_label = "Treatment 2",
  treatment_1_units = NULL,
  treatment_2_units = NULL,
  plot_zero_dose = TRUE,
  contour_color = "gold"
)
}
\arguments{
\item{data}{data.frame with columns \link{dose1, dose2, response}}

\item{treatment_1_label}{used to make default title and axis labels}

\item{treatment_2_label}{used to make default title and axis labels}

\item{treatment_1_units}{used to make default axis labels}

\item{treatment_2_units}{used to make default axis labels}

\item{plot_zero_dose}{plotted on the log scale, zero doses would be
at -Inf, so to show them on the plot, add them as with a slight
separation on the axis.}

\item{contour_color}{default 'gold'}
}
\value{
ggplot2 plot with light-blue to dark-blue tiles the dose
response. Individual plot elements can be over-written and the
plot can be saved with \code{ggplot2::ggsave()}
}
\description{
Plot synergy checkerboard
}
