% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_density_distribution_plots.R
\name{posterior_densities}
\alias{posterior_densities}
\title{Create a plot of the posterior density distributions of modeled parameters
from brmsfit}
\usage{
posterior_densities(
  model,
  predictors_col_name = "predictors",
  half_max_label = NULL,
  title_label = "Posterior Density Plots with Mean and 95\% CI"
)
}
\arguments{
\item{model}{brmsfit.
#'@param predictors_col_name string. Name of the column with the perturbations
that each parameter is being estimated for.}

\item{half_max_label}{string. Label for the half maximal that fits the type of
experiment that was done (i.e. ec50, ic50, ed50, id50, ld50, etc.).}

\item{title_label}{string. Plot title name. (Default = "Posterior Density Plot")}
}
\value{
ggplot object.
}
\description{
The 'basic_stats' function is used to plot the mean, lower confidence
interval, and upper confidence interval
}
