% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-sample-draws.R
\name{posterior_draws_plot}
\alias{posterior_draws_plot}
\title{Plot of Posterior Model Fit Draws}
\usage{
posterior_draws_plot(
  model,
  data = NULL,
  predictors_col_name = "na",
  lower = NULL,
  upper = NULL,
  n = 50,
  facet_var = NULL,
  point_size = 0.75,
  jitter_height = 0,
  jitter_width = 0,
  title = "Dose-Response Posterior Draws",
  xlab = "Log[Molar]",
  ylab = "Response"
)
}
\arguments{
\item{model}{brmsfit model.}

\item{data}{data.frame of data inputted into the model.}

\item{predictors_col_name}{character. The name of the column containing the
predictors for the model fit. If there is only one predictor, then any
character input can be used and will not affect the plot (default = "na").}

\item{lower}{numeric value of the lower bound of `log_dose` to be observed
(default = smallest data$log_dose value > -Inf).}

\item{upper}{numeric value of the upper bound of `log_dose` to be observed
(default = greatest data$log_dose value < Inf).}

\item{n}{numeric value of the number of draws to be observed (default = 50).}

\item{facet_var}{defined variable to determine the facets of the plot. For
models with multiple predictors, include the predictor column name;
otherwise, include the character name of your choosing.}

\item{point_size}{numeric. `geom_jitter` point size (default = 0.75).}

\item{jitter_height}{numeric. the height distance between overlapping points
(default = 0).}

\item{jitter_width}{numeric. the width distance between overlapping points
(default = 0).}

\item{title}{character name for the plot
(default = "Dose-Response Posterior Draws").}

\item{xlab}{character name for the x-axis label (default = "Log[Molar]").}

\item{ylab}{character name for the y-axis label (default = "Response").}
}
\value{
ggplot2::ggplot object.
}
\description{
A plot of a sample of model fit draws from the posterior
distribution from the expected mean and median quantile intervals.
}
\examples{
\dontrun{
  # Consider a model named my_model and data named my_data with a column
  # named predictors containing multiple different perturbations.
  posterior_draws_plot(
    model = my_model,
    data = my_data,
    predictors_col_name = "predictors",
    lower = -12,
    upper = -3,
    n = 50,
    facet_var = predictors,
    jitter_width = 0.10,
    title = "Dose-Response Posterior Draws",
    xlab = "Log[Molar]",
    ylab = "Response")
}

}
