% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_model_draws.R
\name{posterior_response_draws}
\alias{posterior_response_draws}
\title{Data.frame of draw_id, predictors, ec50, hill, top, bottom, and Response from a
sample of the posterior distribution}
\usage{
posterior_response_draws(
  model,
  n = 100,
  lower = -12,
  upper = -3,
  predictors_col_name = "_Intercept"
)
}
\arguments{
\item{model}{brmsfit.}

\item{n}{number of samples from the posterior distribution. (default n = 100)}

\item{lower}{number for the lowest log_dose value to be used for calculating
Response. (default = -12)}

\item{upper}{number for the highest log_dose value to be used for calculating
Response. (default = -3)}

\item{predictors_col_name}{string expression for predictors column in the
input data.frame (default = "_Intercept").}
}
\value{
data.frame.
}
\description{
Data.frame of draw_id, predictors, ec50, hill, top, bottom, and Response from a
sample of the posterior distribution
}
