% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{prepare_init}
\alias{prepare_init}
\title{Helper function to prepare an init for a brms model}
\usage{
prepare_init(init)
}
\arguments{
\item{init}{\code{function} returning an \code{numeric} \code{array} of length \code{1} or a
\code{numeric} value.}
}
\value{
\code{function} returning a \code{numeric} \code{array} of length \code{1}.
}
\description{
Helper function to prepare an init for a brms model
}
