% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{prepare_prior}
\alias{prepare_prior}
\title{Prepare a brmsprior}
\usage{
prepare_prior(prior, ...)
}
\description{
What this gives above brms::prior(...) and friends is
  1) it allow just giving a numeric value rather than constant(<value>)
  2) if it is already a brmsprior it checks that it has specified arguments
}
\details{
This is used in models where want to allow user specified priors but
make sure they are for the right parameters to make sure the model is well
specified.
}
