% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-util.R
\name{prepare_prior}
\alias{prepare_prior}
\title{Prepare a brmsprior}
\usage{
prepare_prior(prior, ...)
}
\arguments{
\item{prior}{\code{\link[brms]{brmsprior}} or \code{numeric}.}

\item{...}{additional arguments to \code{\link[brms]{prior_string}}. If
\code{prior} is a \code{\link[brms]{brmsprior}} then this will check that
the slots have the given values. If prior is numeric, then these arguments
are passed to \code{\link[brms]{prior_string}}}
}
\value{
\code{\link[brms]{brmsprior}}
}
\description{
What this gives above \code{brms::\link[brms]{prior}(...)} and friends is
  1) it allow just giving a numeric value rather than constant(<value>)
  2) if it is already a \code{brms::\link[brms]{brmsprior}} it checks that it
     has specified arguments
}
\details{
This is used in models where want to allow user specified priors but make
sure they are for the right parameters to make sure the model is well
specified.
}
