% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_density_distribution_plots.R
\name{prior_densities}
\alias{prior_densities}
\title{Create a plot of the prior density distributions of modeled parameters
from brmsfit with sample_prior = "only".}
\usage{
prior_densities(
  model,
  predictors_col_name = "predictors",
  half_max_label = NULL,
  title_label = "Prior Density Plots"
)
}
\arguments{
\item{model}{brmsfit of sample_prior = "only".
#'@param predictors_col_name string. Name of the column with the perturbations
that each parameter is being estimated for.}

\item{half_max_label}{string. Label for the half maximal that fits the type of
experiment that was done (i.e. ec50, ic50, ed50, id50, ld50, etc.).}

\item{title_label}{string. Plot title name. (Default = "Prior Density Plot")}
}
\value{
ggplot object.
}
\description{
Create a plot of the prior density distributions of modeled parameters
from brmsfit with sample_prior = "only".
}
