% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-function.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(ec50, hill, top, bottom, log_dose)
}
\arguments{
\item{ec50}{numeric.}

\item{hill}{numeric.}

\item{top}{numeric.}

\item{bottom}{numeric.}

\item{log_dose}{numeric.}
}
\value{
numeric.
}
\description{
A vectorized function of the sigmoid dose-response curve
  equation. This may be need if stan backend is cmdstanr
}
