% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-init.R
\name{sigmoid_agonist_init}
\alias{sigmoid_agonist_init}
\title{Initialize Parameter Values for an Agonist Sigmoid Model}
\usage{
sigmoid_agonist_init(
  ec50 = function() runif(n = 1, min = -7, max = -5),
  hill = function() runif(n = 1, min = 0.8, max = 1.2),
  top = function() runif(n = 1, min = 0.8, max = 1.2),
  bottom = function() runif(n = 1, min = -0.2, max = 0.2),
  ...
)
}
\arguments{
\item{ec50}{\code{numeric} or \code{numeric} returning \code{function} units of the
treatment. If the treatment is \code{log_dose}, the default value of \code{-6}
corresponds \code{1e-6} molar = \code{1} μM}

\item{hill}{\code{numeric} or \code{numeric} returning \code{function} with units
\code{response/log_dose}}

\item{top}{\code{numeric} or \code{numeric} returning \code{function} with units of the
response}

\item{bottom}{\code{numeric} or \code{numeric} returning \code{function} with units of the
response}

\item{...}{additional parameter initialization. Each named argument should be
a function that returns a \code{numeric} or \code{array} depending on the dimension
of the parameter, see \code{\link[=rstan_default_init]{rstan_default_init()}} to use the \pkg{rstan} default
init.}
}
\value{
input for \verb{[brms::brm](init = ...)}
}
\description{
Creating initial values for an agonist sigmoid model parameters
that can be passed to the \code{\link[=sigmoid_model]{sigmoid_model()}} along with the
\code{\link[=sigmoid_agonist_formula]{sigmoid_agonist_formula()}} and \code{\link[=sigmoid_agonist_prior]{sigmoid_agonist_prior()}}.
}
\examples{
\dontrun{
#Consider an activator that has a max response around 50\%, EC50 is estimated
#to be around 1 nM, minimum response is known to be 0.
init <- BayesPharma::sigmoid_agonist_init(
  ec50 = -9,
  top = 0.5)
}
}
\seealso{
\code{\link[=sigmoid_agonist_formula]{sigmoid_agonist_formula()}}, \code{\link[=sigmoid_agonist_prior]{sigmoid_agonist_prior()}}, and
\code{\link[=sigmoid_model]{sigmoid_model()}}
}
