% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-prior.R
\name{sigmoid_agonist_prior}
\alias{sigmoid_agonist_prior}
\title{Create a Prior an Agonist Sigmoid Model}
\usage{
sigmoid_agonist_prior(
  ec50 = brms::prior(prior = normal(-6, 2.5), nlpar = "ec50"),
  hill = brms::prior(prior = normal(1, 1), nlpar = "hill", lb = -0.01),
  top = brms::prior(prior = normal(1, 0.5), nlpar = "top"),
  bottom = brms::prior(prior = normal(0, 0.5), nlpar = "bottom"),
  ...
)
}
\arguments{
\item{ec50}{\link[brms:set_prior]{brms::brmsprior} or \code{numeric}. Prior for the \code{ec50}
parameter. Setting ec50 to a numeric value constrains it to a constant
value. Default: normal(-6, 2.5) where the mean -6 corresponds to a
concentration of 1e-6 or 1 μM.}

\item{hill}{\link[brms:set_prior]{brms::brmsprior} or \code{numeric}. Prior for the \code{hill}
parameter. Default: \code{normal(1, 1)} with and a lower bound of
\code{-0.1}.}

\item{top}{\link[brms:set_prior]{brms::brmsprior} or \code{numeric} giving the prior for the \code{top}
parameter. Setting \code{top} to a \code{numeric} constrains it to a constant
value.}

\item{bottom}{\link[brms:set_prior]{brms::brmsprior} or \code{numeric} giving the prior for the
\code{bottom} parameter. Setting \code{bottom} to \code{numeric} constrains it to a
constant value.}

\item{...}{additional \link[brms:set_prior]{brms::brmsprior} objects.}
}
\value{
\link[brms:set_prior]{brms::brmsprior}
}
\description{
Creates a \code{data.frame} of priors for the sigmoid agonist model

If the function arguments \code{ec50}, \code{hill}, \code{top}, \code{bottom} are \code{NULL}, default
normal distribution priors will be set.

If you would like to set a parameter as a constant, enter a numeric
value for the function argument.

For other distribution options, reference \link[brms:set_prior]{brms::prior} and \href{https://github.com/stan-dev/stan/wiki/Prior-Choice-Recommendations}{Prior Choice Recommendations }
}
\examples{
\dontrun{
# Consider an activator that has a max response around 50\%, EC50 is estimated
# to be around 1 nM, and minimum response is known to be 0.
priors <- sigmoid_agonist_prior(
  ec50 = brms::prior(prior = normal(-9, 0.5), nlpar = "ec50"),
  top = brms::prior(prior = normal(0.5, 0.2), nlpar = "top"),
  bottom = 0)
}
}
\seealso{
\link{sigmoid_agonist_formula}, \link{sigmoid_agonist_init}, and
\link{sigmoid_model}
}
