% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-prior.R
\name{sigmoid_agonist_prior}
\alias{sigmoid_agonist_prior}
\title{Default priors for the sigmoid agoinst model}
\usage{
sigmoid_agonist_prior(
  ec50 = brms::prior(normal(-7, 2.5), nlpar = "ec50"),
  hill = brms::prior(prior = normal(1, 1), nlpar = "hill", ub = 0.01),
  top = brms::prior(normal(100, 25), nlpar = "top"),
  bottom = brms::prior(prior = normal(0, 25), nlpar = "bottom"),
  ...
)
}
\arguments{
\item{ec50}{`brmsprior` or numeric. Prior for the ec50 parameter.
Default: normal(-7, 2.5) where the mean -7 corresponds to a
concentration of 1e-7 or 100 nM. Setting ec50 to a numeric value constrains
it to a constant value.}

\item{hill}{`brmsprior` or numeric. Prior for the hill parameter. Default:
normal(1, 1) with and a lower bound of -0.1.}

\item{top}{`brmsprior` or numeric. Prior for the top parameter.
Default: normal(100, 25). Setting top to a numeric value constrains it to
a constant value.}

\item{bottom}{`brmsprior` or numeric. Prior for the top parameter.
Default: normal(100, 25). Setting bottom to a numeric value constrains it
to a constant value.}

\item{...}{additional `brmsprior` objects.}
}
\value{
`brmsprior` `data.frame`
}
\description{
Creates a data.frame of priors for the sigmoid agonist model

If the function arguments ec50, hill, top, bottom = NULL, default
normal distribution priors will be set.

If you would like to set a parameter as a constant, enter a numeric value for
the function argument.

For other distribution options, reference
<http://mc-stan.org/rstanarm/reference/priors.html#arguments>
}
\examples{
\dontrun{
# Consider an activator that has a max response around 50\%, EC50 is estimated
# to be around 1 uM, and minimum response is known to be 0.
priors <- sigmoid_agonist_prior(
  ec50 = brms::prior(normal(-6, 0.5), nlpar = "ec50"),
  inhibitor = FALSE,
  top = brms::prior(normal(50, 2.5), nlpar = "top"),
  bottom = 0)
}
}
