% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-formula.R
\name{sigmoid_antagonist_formula}
\alias{sigmoid_antagonist_formula}
\title{Create an antagonist sigmoid formula for the brms model}
\usage{
sigmoid_antagonist_formula(predictors = 1, ...)
}
\arguments{
\item{predictors}{Additional formula objects to specify predictors
of non-linear parameters. i.e. what perturbations/experimental
differences should be modeled separately? (Default: 1) should a
random effect be taken into consideration? i.e. cell number,
plate number, etc.}

\item{...}{additional arguments to \code{brms::brmsformula}}
}
\value{
brmsformula
}
\description{
set-up a sigmoid dose response model formula to define
    a non-linear model or multilevel non-linear model for
    \code{ic50}, \code{hill}, \code{top}, and, \code{bottom} for
    use in Bayesian_model and in the BRMS package.
}
\examples{
\dontrun{
  # Data has a string column drug_id with drug identifiers
  # Fit a separate model for each drug
  BayesPharma::sigmoid_antagonist_formula(predictors = 0 + drug_id)

  # Data has a string column plate_id with plate identifiers
  # Estimate the change in response for each plate relative to a global
  # baseline.
  BayesPharma::sigmoid_antagonist_formula(predictors = plate_id)

  # data has columns drug_id and plate_id
  # fit a multilevel model where the drug effect depends on the plate
  BayesPharma::sigmoid_antagonist_formula(
    predictors = 0 + (drug_id|plate_id))
}

}
