% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-init.R
\name{sigmoid_antagonist_init}
\alias{sigmoid_antagonist_init}
\title{Create initialization for the antagonist sigmoid model}
\usage{
sigmoid_antagonist_init(ic50 = -6, hill = -1, top = 1, bottom = 0)
}
\arguments{
\item{ic50}{\code{numeric} or \code{numeric} returning \code{function} units of the
treatment. If the treatment is \code{log(dose)}, the default value of \code{-6}
corresponds 1e-6 molar = 1 μM}

\item{hill}{\code{numeric} or \code{numeric} returning \code{function} with units
response/log_dose}

\item{top}{\code{numeric} or \code{numeric} returning \code{function} in units of the
response}

\item{bottom}{\code{numeric} or \code{numeric} returning \code{funciton} in units of the
response}
}
\value{
input for \verb{[brm][brms::brm](init = ...)}
}
\description{
Creating initial values for the sigmoid model parameters that
can be passed to the \code{\link[=sigmoid_antagonist_model]{sigmoid_antagonist_model()}}
}
\examples{
\dontrun{
#Consider an inhibitor that has a min response around 50\%, IC50 is estimated
#to be around 1 nM, maximum response is known to be around 1,
init <- BayesPharma::sigmoid_antagonist_init(
  ec50 = -9,
  bottom = 0.5)
}
}
\seealso{
\code{\link[=sigma_agonist_formula]{sigma_agonist_formula()}}, \code{\link[=sigma_agonist_prior]{sigma_agonist_prior()}}, and
\code{\link[=sigmoid_agonist_model]{sigmoid_agonist_model()}}
}
