% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-init.R
\name{sigmoid_antagonist_init}
\alias{sigmoid_antagonist_init}
\title{Initialize Parameter Values for an Antagonist Sigmoid Model}
\usage{
sigmoid_antagonist_init(
  ic50 = function() stats::runif(n = 1, min = -7, max = -5),
  hill = function() stats::runif(n = 1, min = -1.2, max = -0.8),
  top = function() stats::runif(n = 1, min = 0.8, max = 1.2),
  bottom = function() stats::runif(n = 1, min = -0.2, max = 0.2),
  ...
)
}
\arguments{
\item{ic50}{\code{numeric} or \code{numeric} returning \code{function} units of the
treatment. If the treatment is \code{log_dose}, the default value of \code{-6}
corresponds \code{1e-6} molar = \code{1} μM}

\item{hill}{\code{numeric} or \code{numeric} returning \code{function} with units
response/log_dose}

\item{top}{\code{numeric} or \code{numeric} returning \code{function} in units of the
response}

\item{bottom}{\code{numeric} or \code{numeric} returning \code{function} in units of the
response}

\item{...}{additional parameter initialization. Each named argument should be
a function that returns an \code{numeric} \code{array} of length \code{1}, see
\code{\link[=rstan_default_init]{rstan_default_init()}} to use the rstan default init.}
}
\value{
input for \verb{[brm][brms::brm](init = ...)}
}
\description{
Creating initial values for an agonist sigmoid model parameters
that can be passed to the \code{\link[=sigmoid_model]{sigmoid_model()}} along with the
\code{\link[=sigmoid_antagonist_formula]{sigmoid_antagonist_formula()}} and \code{\link[=sigmoid_antagonist_prior]{sigmoid_antagonist_prior()}}.
}
\examples{
\dontrun{
# Consider an inhibitor that has a min response around 50\%, IC50 is estimated
# to be around 1 nM, maximum response is known to be around 1,
init <- BayesPharma::sigmoid_antagonist_init(
  ec50 = -9,
  bottom = 0.5)

# By default rstan initializes unspecified initial values to be uniformly at
# random in the range (-2, 2) on the unconstrained scale. For the default
# distributional response, family=gaussian(), the shape parameter sigma' is
# bounded below by zero through this transformation Y = log(X - 0). So, to
# explicitly give the default initialization for 'sigma', we can use
init <- BayesPharma::sigmoid_antagonist_init(
  sigma = BayesPharma::rstan_default_init(lb = 0))
}
}
\seealso{
\code{\link[=sigmoid_antagonist_formula]{sigmoid_antagonist_formula()}}, \code{\link[=sigmoid_antagonist_prior]{sigmoid_antagonist_prior()}}, and
\code{\link[=sigmoid_model]{sigmoid_model()}}
}
