% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sigmoid-priors.R
\name{sigmoid_prior}
\alias{sigmoid_prior}
\title{Default priors for the sigmoid model}
\usage{
sigmoid_prior(
  ec50 = NULL,
  hill = NULL,
  inhibitor = TRUE,
  top = NULL,
  bottom = NULL,
  ...
)
}
\arguments{
\item{ec50}{NULL, numeric units, or brms::prior. NULL will provide a weakly
informative prior with mean of 100 nM. Setting ec50 to a numeric value will
set the ec50 prior to a constant value of the given numeric value. A prior
of choice can be set using brms::prior(). (default = NULL)}

\item{hill}{NULL, numeric value, or brms::prior. NULL will provide a weakly
informative prior. Setting hill to a numeric value will set the hill prior
to a constant value of the given numeric value. A prior of choice can be
set using brms::prior(). (default = NULL)}

\item{inhibitor}{TRUE/FALSE value. This determines if a positive or negative
slope is used for the hill prior. (default = TRUE)}

\item{top}{NULL, numeric value, or `brms::prior`. NULL will provide a weakly
informative prior. Setting top to a numeric value will set the top prior to
a constant value of the given numeric value. A prior of choice can be set
using `brms::prior()`. (default = NULL)}

\item{bottom}{NULL, numeric value, or `brms::prior`. NULL will provide a
weakly
  informative prior. Setting bottom to a numeric value will set the bottom
  prior to a constant value of the given numeric value. A prior of choice can
  be set using `brms::prior()`. (default = NULL)}
}
\value{
brmsprior data.frame
}
\description{
Creates a data.frame of priors for the sigmoid model

If the function arguments ec50, hill, top, bottom = NULL, default
normal distribution priors will be set.

If inhibitor = TRUE, hill is positive. If inhibitor = FALSE, hill is
negative.

If you would like to set a parameter as a constant, enter a numeric value for
the function argument.

For other distribution options, reference
http://mc-stan.org/rstanarm/reference/priors.html#arguments.
}
\examples{
\dontrun{
# Consider an activator that has a max response around 50\%, EC50 is estimated
# to be around 1 uM, and minimum response is known to be 0.
priors <- sigmoid_prior(
  ec50 = brms::prior(normal(-6, 0.5), nlpar = "ec50"),
  hill = NULL,
  inhibitor = FALSE,
  top = brms::prior(normal(50, 2.5), nlpar = "top"),
  bottom = 0)
}
}
