% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-formula.R
\name{tQ_formula}
\alias{tQ_formula}
\title{Define formula for the tQ enzyme kinetics model}
\usage{
tQ_formula(predictors = 1, ...)
}
\arguments{
\item{predictors}{predictors for `kcat` and `kM`.}
}
\description{
Define formula for the tQ enzyme kinetics model
}
