% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tQ_model.R
\name{tQ_formula}
\alias{tQ_formula}
\title{Define formula for the tQ enzyme kinetics model}
\usage{
tQ_formula(multiple_perturbations = FALSE, predictors = 0 + predictors, ...)
}
\arguments{
\item{multiple_perturbations}{(default FALSE)}

\item{predictors}{predictors to use for kcat and kM}
}
\description{
Define formula for the tQ enzyme kinetics model
}
