% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-formula.R
\name{tQ_formula}
\alias{tQ_formula}
\title{Define formula for the tQ enzyme kinetics model}
\usage{
tQ_formula(
  series_index_variable = "series_index",
  treatment_variable = "time",
  treatment_units = "seconds",
  ET_variable = "ET",
  ET_units = "mg/ml",
  ST_variable = "ST",
  ST_units = "mg/ml",
  response_variable = "P",
  response_units = "mg/ml",
  predictors = 1,
  ...
)
}
\arguments{
\item{series_index_variable}{\code{character} variable indexing which
measurements are part of a common time series.}

\item{treatment_variable}{\code{character} variable representing time as the
treatment.}

\item{treatment_units}{\code{character} the units of the treatment variable.}

\item{ET_variable}{\code{character} variable representing the total enzyme
concentration.}

\item{ET_units}{\code{character} variable representing the units of the
enzyme concentration.}

\item{ST_variable}{\code{character} variable representing the total substrate
concentration.}

\item{ST_units}{\code{character} variable representing the units of the
substrate concentration.}

\item{response_variable}{\code{character} variable representing the product.}

\item{response_units}{\code{character} the units of the product.}

\item{predictors}{\code{expression} of predictors for \code{kcat} and
\code{kM} parameters}

\item{...}{additional arguments passed to \code{\link[brms:brmsformula]{brms::brmsformula()}}}
}
\value{
a \code{bpformula}, which is a subclass of
\link[brms:brmsformula]{brms::brmsformula} and can be passed to
\code{\link[=tQ_model]{tQ_model()}}.
}
\description{
Define formula for the tQ enzyme kinetics model
}
\seealso{
\link{tQ_model}, \link{tQ_prior}, \link{tQ_init}, \link{tQ_stanvar} and
\link[brms:brmsformula]{brms::brmsformula}, which this function wraps.
}
