% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-run.R
\name{tQ_init}
\alias{tQ_init}
\title{Initialize parameter values for the the tQ enyzme kinetic model}
\usage{
tQ_init(kcat = 4, kM = 4, chains = 4)
}
\description{
Initialize parameter values for the the tQ enyzme kinetic model
}
