% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-init.R
\name{tQ_init}
\alias{tQ_init}
\title{Initialize parameter values for the tQ enyzme kinetic model}
\usage{
tQ_init(kcat = 4, kM = 4)
}
\arguments{
\item{kcat}{numeric initial value for `r kcat` parameter in the tQ model
(Default 4)}

\item{kM}{numeric initial value for the `r kM` parameter in the tQ model
(default 4)}
}
\value{
input for `r BayesPharma::model_tQ(init = ...)` parameter.
}
\description{
Initialize parameter values for the tQ enyzme kinetic model
}
