% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-run.R
\name{tQ_model}
\alias{tQ_model}
\title{Model for the tQ Enzyme Kinetics Model}
\usage{
tQ_model(
  data,
  formula = tQ_formula(),
  prior = tQ_prior(),
  init = tQ_init(),
  iter = 8000,
  control = list(adapt_delta = 0.99),
  stanvar_function = BayesPharma::tQ_stanvar,
  expose_functions = TRUE,
  ...
)
}
\arguments{
\item{data}{\code{data.frame} of experimental data.
must contain columns \code{response} and any predictors specified in
the formula.}

\item{formula}{\link[brms:brmsformula]{brms::brmsformula} object. To create a dose-response
\link[brms:brmsformula]{brms::brmsformula}, use the \link{tQ_formula} function.}

\item{prior}{\link[brms:set_prior]{brms::brmsprior} for \code{kcat}, and \code{kM}. Use \code{\link[=tQ_formula]{tQ_formula()}} to
create priors to use here.}

\item{init}{\code{list} of \code{lists}, \code{numeric} value, or "random" for the initial
values of the parameters being modeled.}

\item{iter}{\code{numeric} of iterations the model runs. Increasing \code{iter} can
help with model convergence.}

\item{control}{a named \code{list} of parameters to control the sampler's
behavior. Adding \code{max_treedepth} and giving a greater value than \code{10} can
improve model convergence.}

\item{stanvar_function}{stan code for the model}

\item{expose_functions}{\code{logical}. Expose the BayesPharma functions for the
model}

\item{...}{additional arguments passed to \code{\link[brms:brm]{brms::brm()}}.}
}
\value{
\code{bpfit} object, which is a wrapper around a \link[brms:brmsfit-class]{brms::brmsfit} object.
}
\description{
Model for the tQ Enzyme Kinetics Model
}
\references{
Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation:
Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7,
17018 (2017). https://doi.org/10.1038/s41598-017-17072-z
}
\seealso{
\link{tQ_formula}, \link{tQ_prior}, \link{tQ_init}, or \link{tQ_stanvar}
}
