% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tQ_model.R
\name{tQ_prior}
\alias{tQ_prior}
\title{Define priors for the tQ enzyme kinetics model}
\usage{
tQ_prior(kcat = NULL, kM = NULL)
}
\arguments{
\item{kcat}{prior for kcat parameter (Default: NULL). Given a numeric value,
it will be used as a constant and not estimated.}

\item{kM}{prior for kM parameter (Default: NULL). Given a numeric value,
it will be used as a constant and not estimated.}
}
\description{
Default priors are gamma(4, 1) for both `kcat` and `kM`. We use the gamma
distribution because it is naturally lower bounded by 0. The first parameter
is the shape, `alpha=4`, and the second is the rate, `beta=1`. The mean of
gamma distributions is `alpha/beta` and the variance is `alpha/beta^2`.
}
