% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-stanvars.R
\docType{data}
\name{tQ_stanvar}
\alias{tQ_stanvar}
\title{stanvar for the tQ enzyme kinetics model}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
tQ_stanvar
}
\description{
stanvar for the tQ enzyme kinetics model
}
\keyword{datasets}
