% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-stanvars.R
\docType{data}
\name{tQ_stanvar}
\alias{tQ_stanvar}
\title{Stan code for the tQ enzyme kinetics model}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
tQ_stanvar
}
\description{
The tQ is an ordinary differential equation model for the total
quasi-steady-state assumption kinetics defined in
(Choi et al., 2017), which is related to the Michaelis-Menten kinetics model,
but doesn't assume the enzyme concentration is negligibly small.
}
\details{
To implement it in \pkg{Stan}, the ODE function \code{tQ_ode} is defined
}
\keyword{datasets}
