% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tQ-stanvars.R
\name{tQ_stanvar}
\alias{tQ_stanvar}
\title{Stan Code for the tQ Enzyme Kinetics Model}
\usage{
tQ_stanvar()
}
\description{
The tQ is an ordinary differential equation model for the total
quasi-steady-state assumption kinetics defined in (Choi et al., 2017), which
is related to the Michaelis-Menten kinetics model, but doesn't assume the
enzyme concentration is negligibly small.

To implement the tQ model in \pkg{Stan}, the function \code{tQ_ode} is defined
and then passed to \code{tQ_single} to integrate it using the \emph{stiff backward
differentiation formula (BDF) method}. To fit multiple time series
in one model, the \code{tQ_multiple} can be used. Note that to handle fitting
time-series with different numbers of observations, an additional
\code{series_index} argument is used. Note that observations in the same
time-series should be in sequential order in the supplied data.
}
\examples{
\dontrun{
brms::brm(
  data = ...,
  formula = brms::brmsformula(
    P ~ tQ_multiple(series_index, time, kcat, kM, ET, ST),
    kcat + kM ~ 1,
    nl = TRUE,
    loop = FALSE),
  prior = ...,
  init =  ...,
  stanvars = BayesPharma::tQ_stanvar())
}
}
\references{
Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation:
Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7,
17018 (2017). https://doi.org/10.1038/s41598-017-17072-z
}
\seealso{
\link{tQ_model},
}
