% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{traceplot}
\alias{traceplot}
\title{Create a plot of MCMC chains and sample value at each MCMC iteration}
\usage{
traceplot(model, title = NULL)
}
\arguments{
\item{model}{brmsfit}

\item{title}{string for the plot title (default = NULL)}
}
\value{
ggplot object
}
\description{
Create a plot of MCMC chains and sample value at each MCMC iteration
}
